\name{eco}
\alias{eco}
\docType{data}
\title{
ADAC Eco test data
}
\description{
ADAC Ecotest data with clusterings.
}
\usage{data("eco")}
\format{
  A data frame with 753 observations on the following 21 variables.
  \describe{
    \item{\code{V1}}{a factor with levels \code{ Hersteller} \code{1} \code{10} \code{100} \code{101} \code{102} \code{103} \code{104} \code{105} \code{106} \code{107} \code{108} \code{109} \code{11} \code{110} \code{111} \code{112} \code{113} \code{114} \code{115} \code{116} \code{117} \code{118} \code{119} \code{12} \code{120} \code{121} \code{122} \code{123} \code{124} \code{125} \code{126} \code{127} \code{128} \code{129} \code{13} \code{130} \code{131} \code{132} \code{133} \code{134} \code{135} \code{136} \code{137} \code{138} \code{139} \code{14} \code{140} \code{141} \code{142} \code{143} \code{144} \code{145} \code{146} \code{147} \code{148} \code{149} \code{15} \code{150} \code{151} \code{152} \code{153} \code{154} \code{155} \code{156} \code{157} \code{158} \code{159} \code{16} \code{160} \code{161} \code{162} \code{163} \code{164} \code{165} \code{166} \code{167} \code{168} \code{169} \code{17} \code{170} \code{171} \code{172} \code{173} \code{174} \code{175} \code{176} \code{177} \code{178} \code{179} \code{18} \code{180} \code{181} \code{182} \code{183} \code{184} \code{185} \code{186} \code{187} \code{188} \code{189} \code{19} \code{190} \code{191} \code{192} \code{193} \code{194} \code{195} \code{196} \code{197} \code{198} \code{199} \code{2} \code{20} \code{200} \code{201} \code{202} \code{203} \code{204} \code{205} \code{206} \code{207} \code{208} \code{209} \code{21} \code{210} \code{211} \code{212} \code{213} \code{214} \code{215} \code{216} \code{217} \code{218} \code{219} \code{22} \code{220} \code{221} \code{222} \code{223} \code{224} \code{225} \code{226} \code{227} \code{228} \code{229} \code{23} \code{230} \code{231} \code{232} \code{233} \code{234} \code{235} \code{236} \code{237} \code{238} \code{239} \code{24} \code{240} \code{241} \code{242} \code{243} \code{244} \code{245} \code{246} \code{247} \code{248} \code{249} \code{25} \code{250} \code{251} \code{252} \code{253} \code{254} \code{255} \code{256} \code{257} \code{258} \code{259} \code{26} \code{260} \code{261} \code{262} \code{263} \code{264} \code{265} \code{266} \code{267} \code{268} \code{269} \code{27} \code{270} \code{271} \code{272} \code{273} \code{274} \code{275} \code{276} \code{277} \code{278} \code{279} \code{28} \code{280} \code{281} \code{282} \code{283} \code{284} \code{285} \code{286} \code{287} \code{288} \code{289} \code{29} \code{290} \code{291} \code{292} \code{293} \code{294} \code{295} \code{296} \code{297} \code{298} \code{299} \code{3} \code{30} \code{300} \code{301} \code{302} \code{303} \code{304} \code{305} \code{306} \code{307} \code{308} \code{309} \code{31} \code{310} \code{311} \code{312} \code{313} \code{314} \code{315} \code{316} \code{317} \code{318} \code{319} \code{32} \code{320} \code{321} \code{322} \code{323} \code{324} \code{325} \code{326} \code{327} \code{328} \code{329} \code{33} \code{330} \code{331} \code{332} \code{333} \code{334} \code{335} \code{336} \code{337} \code{338} \code{339} \code{34} \code{340} \code{341} \code{342} \code{343} \code{344} \code{345} \code{346} \code{347} \code{348} \code{349} \code{35} \code{350} \code{351} \code{352} \code{353} \code{354} \code{355} \code{356} \code{357} \code{358} \code{359} \code{36} \code{360} \code{361} \code{362} \code{363} \code{364} \code{365} \code{366} \code{367} \code{368} \code{369} \code{37} \code{370} \code{371} \code{372} \code{373} \code{374} \code{375} \code{376} \code{377} \code{378} \code{379} \code{38} \code{380} \code{381} \code{382} \code{383} \code{384} \code{385} \code{386} \code{387} \code{388} \code{389} \code{39} \code{390} \code{391} \code{392} \code{393} \code{394} \code{395} \code{396} \code{397} \code{398} \code{399} \code{4} \code{40} \code{400} \code{401} \code{402} \code{403} \code{404} \code{405} \code{406} \code{407} \code{408} \code{409} \code{41} \code{410} \code{411} \code{412} \code{413} \code{414} \code{415} \code{416} \code{417} \code{418} \code{419} \code{42} \code{420} \code{421} \code{422} \code{423} \code{424} \code{425} \code{426} \code{427} \code{428} \code{429} \code{43} \code{430} \code{431} \code{432} \code{433} \code{434} \code{435} \code{436} \code{437} \code{438} \code{439} \code{44} \code{440} \code{441} \code{442} \code{443} \code{444} \code{445} \code{446} \code{447} \code{448} \code{449} \code{45} \code{450} \code{451} \code{452} \code{453} \code{454} \code{455} \code{456} \code{457} \code{458} \code{459} \code{46} \code{460} \code{461} \code{462} \code{463} \code{464} \code{465} \code{466} \code{467} \code{468} \code{469} \code{47} \code{470} \code{471} \code{472} \code{473} \code{474} \code{475} \code{476} \code{477} \code{478} \code{479} \code{48} \code{480} \code{481} \code{482} \code{483} \code{484} \code{485} \code{486} \code{487} \code{488} \code{489} \code{49} \code{490} \code{491} \code{492} \code{493} \code{494} \code{495} \code{496} \code{497} \code{498} \code{499} \code{5} \code{50} \code{500} \code{501} \code{502} \code{503} \code{504} \code{505} \code{506} \code{507} \code{508} \code{509} \code{51} \code{510} \code{511} \code{512} \code{513} \code{514} \code{515} \code{516} \code{517} \code{518} \code{519} \code{52} \code{520} \code{521} \code{522} \code{523} \code{524} \code{525} \code{526} \code{527} \code{528} \code{529} \code{53} \code{530} \code{531} \code{532} \code{533} \code{534} \code{535} \code{536} \code{537} \code{538} \code{539} \code{54} \code{540} \code{541} \code{542} \code{543} \code{544} \code{545} \code{546} \code{547} \code{548} \code{549} \code{55} \code{550} \code{551} \code{552} \code{553} \code{554} \code{555} \code{556} \code{557} \code{558} \code{559} \code{56} \code{560} \code{561} \code{562} \code{563} \code{564} \code{565} \code{566} \code{567} \code{568} \code{569} \code{57} \code{570} \code{571} \code{572} \code{573} \code{574} \code{575} \code{576} \code{577} \code{578} \code{579} \code{58} \code{580} \code{581} \code{582} \code{583} \code{584} \code{585} \code{586} \code{587} \code{588} \code{589} \code{59} \code{590} \code{591} \code{592} \code{593} \code{594} \code{595} \code{596} \code{597} \code{598} \code{599} \code{6} \code{60} \code{600} \code{601} \code{602} \code{603} \code{604} \code{605} \code{606} \code{607} \code{608} \code{609} \code{61} \code{610} \code{611} \code{612} \code{613} \code{614} \code{615} \code{616} \code{617} \code{618} \code{619} \code{62} \code{620} \code{621} \code{622} \code{623} \code{624} \code{625} \code{626} \code{627} \code{628} \code{629} \code{63} \code{630} \code{631} \code{632} \code{633} \code{634} \code{635} \code{636} \code{637} \code{638} \code{639} \code{64} \code{640} \code{641} \code{642} \code{643} \code{644} \code{645} \code{646} \code{647} \code{648} \code{649} \code{65} \code{650} \code{651} \code{652} \code{653} \code{654} \code{655} \code{656} \code{657} \code{658} \code{659} \code{66} \code{660} \code{661} \code{662} \code{663} \code{664} \code{665} \code{666} \code{667} \code{668} \code{669} \code{67} \code{670} \code{671} \code{672} \code{673} \code{674} \code{675} \code{676} \code{677} \code{678} \code{679} \code{68} \code{680} \code{681} \code{682} \code{683} \code{684} \code{685} \code{686} \code{687} \code{688} \code{689} \code{69} \code{690} \code{691} \code{692} \code{693} \code{694} \code{695} \code{696} \code{697} \code{698} \code{699} \code{7} \code{70} \code{700} \code{701} \code{702} \code{703} \code{704} \code{705} \code{706} \code{707} \code{708} \code{709} \code{71} \code{710} \code{711} \code{712} \code{713} \code{714} \code{715} \code{716} \code{717} \code{718} \code{719} \code{72} \code{720} \code{721} \code{722} \code{723} \code{724} \code{725} \code{726} \code{727} \code{728} \code{729} \code{73} \code{730} \code{731} \code{732} \code{733} \code{734} \code{735} \code{736} \code{737} \code{738} \code{739} \code{74} \code{740} \code{741} \code{742} \code{743} \code{744} \code{745} \code{746} \code{747} \code{748} \code{749} \code{75} \code{750} \code{751} \code{752} \code{76} \code{77} \code{78} \code{79} \code{8} \code{80} \code{81} \code{82} \code{83} \code{84} \code{85} \code{86} \code{87} \code{88} \code{89} \code{9} \code{90} \code{91} \code{92} \code{93} \code{94} \code{95} \code{96} \code{97} \code{98} \code{99}}
    \item{\code{V2}}{a factor with levels \code{Alfa Romeo} \code{Audi} \code{BMW} \code{BMW Alpina} \code{Brilliance} \code{Cadillac} \code{Chevrolet} \code{Chevrolet (EU)} \code{Chrysler} \code{Citroen} \code{Dacia} \code{Daewoo} \code{Daihatsu} \code{Dodge} \code{Fiat} \code{Ford} \code{Honda} \code{Hyundai} \code{Jaguar} \code{Jiangling} \code{Kia} \code{KIA} \code{Lada} \code{Lancia} \code{Land Rover} \code{Lexus} \code{Mazda} \code{Mercedes} \code{MG} \code{Mini} \code{Mitsubishi} \code{Modell} \code{Nissan} \code{Opel} \code{Peugeot} \code{Porsche} \code{Renault} \code{Rover} \code{Saab} \code{Seat} \code{Skoda} \code{smart} \code{SsangYong} \code{Subaru} \code{Suzuki} \code{Toyota} \code{Volvo} \code{VW}}
    \item{\code{V3}}{a factor with levels \code{Alfa Romeo 147 1.9 JTD 16V M-Jet Distinctive} \code{Alfa Romeo 159 1.9 JTDM 16V Distinctive (DPF)} \code{Alfa Romeo 159 Sportwagon 1.9 JTDM 16V Distinctive (DPF)} \code{Alfa Romeo 166 2.4 JTD 20V Multijet Distinctive} \code{Alfa Romeo Brera 2.2 JTS 16V Skyview} \code{Alfa Romeo Brera 2.4 JTDM 20V Q-Tronic (DPF)} \code{Alfa Romeo GT 1.9 JTD 16V Multijet Progression} \code{Alfa Romeo GT 1.9 JTDM 16V Q2 Progression (DPF)} \code{Alfa Romeo GT 2.0 16V JTS 16V Distinctive} \code{Alfa Romeo Spider 2.2 JTS 16V Exclusive} \code{Alfa Romeo Spider 2.4 JTDM 20V Exclusive (DPF)} \code{Audi A2 1.4 TDI} \code{Audi A3 1.6 Attraction} \code{Audi A3 1.6 FSI Ambiente} \code{Audi A3 1.9 TDI Ambition (DPF)} \code{Audi A3 1.9 TDI Attraction} \code{Audi A3 1.9 TDI e Ambition (DPF)} \code{Audi A3 2.0 FSI Ambition} \code{Audi A3 2.0 TDI Attraction} \code{Audi A3 Cabriolet 1.8 TFSI Ambition} \code{Audi A3 Sportback 1.8 TFSI Ambition} \code{Audi A3 Sportback 2.0 TDI Attraction} \code{Audi A3 Sportback 2.0 TFSI Ambition S tronic (DSG)} \code{Audi A4 1.8 TFSI Ambition} \code{Audi A4 2.0 TDI} \code{Audi A4 2.0 TDI (DPF)} \code{Audi A4 2.0 TDI Ambition (DPF)} \code{Audi A4 2.0 TFSI} \code{Audi A4 2.7 TDI Ambition multitronic (DPF)} \code{Audi A4 3.0 TDI Ambition quattro (DPF)} \code{Audi A4 3.2 FSI Ambition quattro tiptronic} \code{Audi A4 Avant 1.8 TFSI Ambition} \code{Audi A4 Avant 2.0 TDI} \code{Audi A4 Avant 3.0} \code{Audi A4 Avant 3.0 multitronic} \code{Audi A5 3.0 TDI quattro (DPF)} \code{Audi A5 3.2 FSI multitronic} \code{Audi A6 2.0 TDI (DPF)} \code{Audi A6 2.4} \code{Audi A8 3.0 TDI quattro tiptronic (DPF)} \code{Audi A8 4.2 TDI quattro tiptronic (DPF)} \code{Audi Q7 3.0 TDI quattro tiptronic (DPF)} \code{Audi R8} \code{Audi S3} \code{Audi TT Roadster 1.8 T tiptronic} \code{Audi TT Roadster 2.0 TFSI} \code{BMW 116i} \code{BMW 118d} \code{BMW 118d (DPF)} \code{BMW 120d} \code{BMW 120d (DPF)} \code{BMW 120d Cabriolet (DPF)} \code{BMW 120i} \code{BMW 123d (DPF)} \code{BMW 123d Coup<e9> (DPF)} \code{BMW 125i Cabriolet} \code{BMW 130i} \code{BMW 135i Coup<e9>} \code{BMW 318d (DPF)} \code{BMW 318d touring (DPF)} \code{BMW 318i} \code{BMW 318ti compact Edition Lifestyle} \code{BMW 320Cd Cabriolet} \code{BMW 320Ci Cabriolet} \code{BMW 320d (DPF)} \code{BMW 320d touring (DPF)} \code{BMW 320i touring} \code{BMW 325Ci Cabriolet Steptronic} \code{BMW 325d (DPF)} \code{BMW 330Cd Coup<e9>} \code{BMW 330d Cabriolet Steptronic (DPF)} \code{BMW 330d Steptronic (DPF)} \code{BMW 330i} \code{BMW 330i Cabriolet} \code{BMW 330i SMG} \code{BMW 330i Steptronic} \code{BMW 330xi touring Steptronic} \code{BMW 335d Coup<e9> Steptronic (DPF)} \code{BMW 520d (DPF)} \code{BMW 520i Steptronic} \code{BMW 525d (DPF)} \code{BMW 530d} \code{BMW 530d touring Steptronic (DPF)} \code{BMW 530i} \code{BMW 535d Steptronic (DPF)} \code{BMW 630Ci Coup<e9> Steptronic} \code{BMW 645Ci Cabriolet} \code{BMW 730d Steptronic (DPF)} \code{BMW 745d Steptronic (DPF)} \code{BMW Alpina D3 (DPF)} \code{BMW M3 Cabriolet M DKG} \code{BMW M3 Coup<e9>} \code{BMW M5 touring SMG} \code{BMW X3 2.0d Steptronic (DPF)} \code{BMW X3 2.5si Steptronic} \code{BMW X3 3.0d Automatik} \code{BMW X3 3.0sd Steptronic (DPF)} \code{BMW X5 3.0d Steptronic} \code{BMW X5 3.0d Steptronic (DPF)} \code{BMW X5 4.4i Steptronic} \code{BMW X6 xDrive35d Sport-Automatic (DPF)} \code{BMW Z4 Roadster 2.2i} \code{BMW Z4 Roadster 3.0i Steptronic} \code{Brilliance BS6 2.0 Deluxe} \code{Cadillac BLS Wagon 1.9 TiD Sport Automatik (DPF)} \code{Cadillac CTS 2.6 V6 Elegance Automatik} \code{Chevrolet (EU) Aveo 1.2 LS} \code{Chevrolet (EU) Captiva 2.0 D LT Sport 4WD (7-Sitzer) (DPF)} \code{Chevrolet (EU) Captiva 3.2 LT 4WD Automatik (7-Sitzer)} \code{Chevrolet (EU) Epica 2.5 LT Automatik} \code{Chevrolet (EU) HHR 2.4 LT} \code{Chevrolet (EU) Kalos 1.4 16V SX} \code{Chevrolet (EU) Matiz 0.8 LPG S (Autogasbetrieb)} \code{Chevrolet (EU) Matiz 0.8 LPG S (Benzinbetrieb)} \code{Chevrolet (EU) Matiz 1.0 SX} \code{Chevrolet (EU) Nubira 1.6 SX} \code{Chevrolet (EU) Nubira Wagon 1.8 LPG CDX (Autogasbetrieb) (LPG)} \code{Chevrolet (EU) Nubira Wagon 1.8 LPG CDX (Benzinbetrieb)} \code{Chevrolet (EU) Nubira Wagon 2.0 D CDX (DPF)} \code{Chevrolet (EU) Rezzo 2.0 LPG CDX (Autogasbetrieb)} \code{Chevrolet (EU) Rezzo 2.0 LPG CDX (Benzinbetrieb)} \code{Chrysler 300C 5.7 V8 Automatik} \code{Chrysler Crossfire 3.2 V6 Automatik} \code{Chrysler Crossfire Roadster 3.2 V6 Automatik} \code{Chrysler PT Cruiser Cabrio 2.4 Limited} \code{Chrysler Sebring 2.0 CRD Limited} \code{Chrysler Sebring Cabrio 2.7 Limited Automatik} \code{Chrysler Sebring Cabrio 2.7 LX Automatik} \code{Citroen Berlingo Kombi 1.4 Bivalent Multispace Plus (Benzinbetrieb)} \code{Citroen Berlingo Kombi 1.4 Bivalent Multispace Plus (Erdgasbetrieb)} \code{Citroen C-Crosser 2.2 HDi FAP Exclusive} \code{Citroen C1 1.0 Style} \code{Citroen C1 HDi 55 Style} \code{Citroen C2 1.4 16V StopStart SensoDrive} \code{Citroen C2 1.4 VTR SensoDrive} \code{Citroen C3 1.4 16V Stop Start SensoDrive} \code{Citroen C3 HDi 110 FAP Exclusive} \code{Citroen C3 HDi 90 Exclusive} \code{Citroen C3 Pluriel 1.6 16V Exclusive SensoDrive} \code{Citroen C4 1.6 16V Confort} \code{Citroen C4 Coup<e9> 1.6 16V VTR} \code{Citroen C4 Coup<e9> HDi 135 FAP VTR Plus Automatik} \code{Citroen C4 HDi 110 FAP Confort} \code{Citroen C4 Picasso 1.8 16V Tendance} \code{Citroen C4 Picasso HDi 110 FAP Tendance EGS6} \code{Citroen C5 2.0 16V Confort} \code{Citroen C5 HDi 110 FAP Style} \code{Citroen C5 HDi 135 FAP Confort Automatik} \code{Citroen C5 HDi 135 FAP Exclusive} \code{Citroen C5 Kombi 1.8 16V Tendance} \code{Citroen C5 Kombi 2.0 16V Tendance} \code{Citroen C5 Kombi HDi 110 FAP Tendance} \code{Citroen C5 Tourer HDi 110 FAP Confort} \code{Citroen C6 HDi 170 Biturbo FAP Pallas} \code{Citroen C6 V6 HDi 205 Biturbo FAP Exclusive Automatik} \code{Citroen C8 2.0 16V Tendance} \code{Citroen Grand C4 Picasso 1.8 16V Tendance} \code{Citroen Grand C4 Picasso HDi 135 FAP Exclusive EGS6} \code{Citroen Jumpy Kombi HDi 135 FAP Club lang (8-Sitzer)} \code{Citroen Xsara Kombi 1.4 HDi SX} \code{Citroen Xsara Picasso 2.0 16V Exclusive Automatik} \code{Corvette C5 Cabrio} \code{Dacia Logan 1.4 Ambiance} \code{Dacia Logan 1.5 dCi Laur<e9>ate} \code{Dacia Logan 1.6 Laur<e9>ate} \code{Dacia Logan MCV 1.5 dCi Laur<e9>ate} \code{Daewoo Evanda 2.0 CDX} \code{Daewoo Lacetti 1.8 CDX} \code{Daewoo Matiz 1.0 SE} \code{Daewoo Nubira 1.8 CDX} \code{Daewoo Nubira Wagon 1.6 SX} \code{Daihatsu Copen} \code{Daihatsu Cuore 1.0 Top} \code{Daihatsu Materia 1.5} \code{Daihatsu Sirion 1.3} \code{Daihatsu Trevis 1.0} \code{Dodge Caliber 2.0 CRD SXT (DPF offen)} \code{Dodge Nitro 2.8 CRD SXT 4WD Automatik (DPF)} \code{Fiat 500 1.3 JTD Multijet 16V Lounge (DPF)} \code{Fiat 500 1.4 16V Lounge} \code{Fiat 500 1.4 16V Sport} \code{Fiat Bravo 1.4 T-Jet 16V Dynamic} \code{Fiat Bravo 1.9 JTD Multijet 8V Emotion (DPF)} \code{Fiat Croma 1.9 JTD Multijet 16V Emotion Automatik (DPF)} \code{Fiat Dobl<f2> Kombi 1.9 JTD Multijet 8V Dynamic (DPF)} \code{Fiat Grande Punto 1.2 8V Dynamic} \code{Fiat Grande Punto 1.3 JTD Multijet 16V Dynamic} \code{Fiat Idea 1.3 JTD 70 Multijet 16V Dynamic} \code{Fiat Idea 1.9 JTD 100 Multijet 8V Emotion} \code{Fiat Linea 1.4 T-Jet 16V Emotion} \code{Fiat Multipla 1.9 JTD Dynamic} \code{Fiat Multipla 1.9 JTD Multijet 8V Dynamic (DPF)} \code{Fiat Panda 1.1 8V Active} \code{Fiat Panda 1.2 8V Emotion} \code{Fiat Panda 1.2 8V Natural Power Panda Panda (Benzinbetrieb)} \code{Fiat Panda 1.2 8V Natural Power Panda Panda (Erdgasbetrieb)} \code{Fiat Panda 1.3 JTD Multijet 16V Cross 4x4} \code{Fiat Panda 1.3 JTD Multijet 16V Emotion} \code{Fiat Panda 1.4 16V 100 HP} \code{Fiat Punto 1.3 JTD 70 Multijet Dynamic} \code{Fiat Scudo Kombi 140 Multijet DPF Panorama Executive lang} \code{Fiat Stilo 1.2 16V Active} \code{Fiat Stilo 1.9 JTD Multijet 16V Dynamic} \code{Ford C-MAX 2.0 TDCi DPF Titanium} \code{Ford C-MAX 2.0 Titanium} \code{Ford Fiesta 1.4 TDCi Ghia} \code{Ford Fiesta 1.6 Trend} \code{Ford Fiesta ST} \code{Ford Focus 1.4} \code{Ford Focus 1.6 Futura} \code{Ford Focus 1.6 TDCi DPF Ghia} \code{Ford Focus 1.6 Ti-VCT Trend} \code{Ford Focus 1.8 FFV Ambiente (Ethanol-Betrieb)} \code{Ford Focus 1.8 Style} \code{Ford Focus 2.0 Sport} \code{Ford Focus 2.0 TDCi DPF Titanium} \code{Ford Focus C-MAX 1.8 Trend} \code{Ford Focus C-MAX 2.0 TDCi DPF Ghia} \code{Ford Focus C-MAX 2.0 TDCi DPF Trend} \code{Ford Focus C-MAX 2.0 TDCi Trend} \code{Ford Focus Coup<e9>-Cabriolet 2.0 TDCi DPF Titanium} \code{Ford Focus Coup<e9>-Cabriolet 2.0 Titanium} \code{Ford Focus Turnier 1.6} \code{Ford Focus Turnier 1.8 Ghia} \code{Ford Focus Turnier 1.8 TDCi Futura} \code{Ford Focus Turnier 2.0 TDCi DPF Titanium} \code{Ford Galaxy 2.0 Trend} \code{Ford Kuga 2.0 TDCi DPF Titanium 4x4} \code{Ford Mondeo 1.6 Trend} \code{Ford Mondeo 1.8 Futura} \code{Ford Mondeo 1.8 SCi Ambiente} \code{Ford Mondeo 2.0 TDCi DPF Ghia X} \code{Ford Mondeo Turnier 2.0 TDCi DPF Titanium} \code{Ford Mondeo Turnier 2.0 TDCi Trend (5-Gang)} \code{Ford Mondeo Turnier 2.0 Trend} \code{Ford Mondeo Turnier 2.5 Turbo Titanium X} \code{Ford Ranger 2.5 TDCi Doppelkabine XLT Limited} \code{Ford S-MAX 2.5 Trend} \code{Ford Streetka 1.6 8V Elegance} \code{Ford Tourneo Connect Kombi 1.8 TDCi kurz} \code{Honda Accord 2.0i Executive} \code{Honda Accord 2.4 Executive Automatik} \code{Honda Accord Tourer 2.2 i-CTDi Sport} \code{Honda Civic 1.3 DSi IMA} \code{Honda Civic 1.3 i-DSi i-VTEC IMA} \code{Honda Civic 1.4i LS} \code{Honda Civic 1.6i ES} \code{Honda Civic 2.2 i-CTDi Type S (DPF)} \code{Honda CR-V 2.0i ES Automatik} \code{Honda FR-V 1.8 Executive Automatik} \code{Honda FR-V 2.0i Executive} \code{Honda FR-V 2.2 i-CTDi Executive (DPF)} \code{Honda FR-V 2.2i-CTDi Comfort} \code{Honda Jazz 1.4 ES Sport} \code{Honda Legend 3.5 V6 Automatik} \code{Honda S2000} \code{Honda Stream 2.0i Sport} \code{Hyundai Atos 1.1 ComfortVersion} \code{Hyundai Coup<e9> 2.0 GLS} \code{Hyundai Elantra 2.0 GLS} \code{Hyundai Getz 1.1 GL} \code{Hyundai Getz 1.4 GLS} \code{Hyundai Getz 1.5 CRDi GLS} \code{Hyundai Grandeur 2.2 CRDi Automatik (DPF)} \code{Hyundai Grandeur 3.3 V6 Automatik} \code{Hyundai i10 1.1 Classic E} \code{Hyundai i10 1.1 CRDi Style} \code{Hyundai i30 1.4 Comfort} \code{Hyundai i30cw 1.6 Comfort} \code{Hyundai i30cw 1.6 CRDi Comfort (DPF)} \code{Hyundai Matrix 1.5 CRDi VGT GLS} \code{Hyundai Sonata 2.4 GLS Automatik} \code{Hyundai Terracan 2.9 CRDi GLS Automatik} \code{Hyundai Trajet 2.0 GLS} \code{Hyundai Tucson 2.0 CRDi GLS 4WD} \code{Hyundai XG 350 Automatik} \code{Jaguar XJ6 3.0 V6 Automatik} \code{Jaguar XKR Coup<e9> Automatik} \code{Jiangling Landwind 2.4 SC2 4WD} \code{KIA Carens 2.0 CWT EX} \code{KIA Carnival 2.9 CRDi EX (DPF offen)} \code{KIA Carnival II 2.9 CRDi EX} \code{KIA cee<b4>d 1.6 CRDi 115 EX (DPF)} \code{KIA cee<b4>d 1.6 EX} \code{KIA cee<b4>d Sporty Wagon 1.6 CRDi 115 LX (DPF)} \code{KIA Cerato 1.6 EX} \code{KIA Cerato 2.0 CRDi EX} \code{KIA Cerato 2.0 EX} \code{KIA Magentis 2.0 CRDi EX (DPF)} \code{KIA Opirus 3.8 Automatik} \code{KIA Picanto 1.1 Cool} \code{KIA Picanto 1.1 EX} \code{KIA pro_cee<b4>d 2.0 CRDi TX (DPF)} \code{KIA Rio 1.4 EX} \code{KIA Rio 1.5 CRDi EX Top} \code{KIA Rio 1.5 LS} \code{KIA Rio 1.6 EX Top} \code{KIA Shuma II 1.8 Active} \code{KIA Sorento 3.5 EX Automatik} \code{Klasse} \code{Lada 1118 Kalina 1.6 8V} \code{Lancia Musa 1.9 jtd Multijet 8v Oro} \code{Lancia Thesis 2.4 jtd Multijet 20v Emblema} \code{Lancia Ypsilon 1.3 jtd Multijet 16v Argento D.F.N-System} \code{Lancia Ypsilon 1.3 jtd Multijet 16v Platino} \code{Lancia Ypsilon 1.4 16v Oro} \code{Land Rover Freelander Td4 SE (DPF)} \code{Lexus GS 430 Luxury Automatik} \code{Lexus GS 460 Luxury Automatik} \code{Lexus IS 220d Luxury (DPF)} \code{Lexus IS 250 Luxury Automatik} \code{Lexus LS 460 Impression Automatik} \code{Lexus LS 600h L Ambience Wellness Automatik} \code{Lexus RX 300 Executive Automatik} \code{Lexus RX 350 Executive Automatik} \code{Lexus RX 400h Luxury Automatik} \code{Lexus SC 430 Automatik} \code{Mazda 2 1.3 Impression} \code{Mazda 2 1.4 Active} \code{Mazda 2 1.4 CD Active} \code{Mazda 2 1.4 CD Independence} \code{Mazda 2 1.4 Exclusive} \code{Mazda 2 1.5 Impression} \code{Mazda 2 1.6 Top} \code{Mazda 3 2.0 Top} \code{Mazda 3 Sport 1.4 Comfort} \code{Mazda 3 Sport 1.6 Exclusive} \code{Mazda 3 Sport 1.6 Exclusive Activematik} \code{Mazda 3 Sport 2.0 CD DPF Active} \code{Mazda 5 1.8 Exclusive} \code{Mazda 5 2.0 CD DPF Top} \code{Mazda 5 2.0 Top} \code{Mazda 6 1.8 Comfort} \code{Mazda 6 1.8 Exclusive} \code{Mazda 6 2.0 CD DPF Active} \code{Mazda 6 Sport 2.0 Exclusive} \code{Mazda 6 Sport 2.3 Top} \code{Mazda 6 Sport Kombi 2.0 CD Top (DPF)} \code{Mazda 6 Sport Kombi 2.0 Exclusive} \code{Mazda BT-50 XL-Cab Toplands} \code{Mazda CX-7 2.3 DISI Expression} \code{Mazda MPV 2.0 TD Comfort6} \code{Mazda MX-5 1.9} \code{Mazda MX-5 Roadster-Coupe 2.0 Expression} \code{Mazda RX-8 High-Power Revolution} \code{Mazda RX-8 STD-Power Revolution} \code{Mercedes A 160 CDI Classic} \code{Mercedes A 160 LR Elegance} \code{Mercedes A 170 Classic} \code{Mercedes A 180 CDI Avantgarde (DPF)} \code{Mercedes A 200 Classic Autotronic} \code{Mercedes B 150} \code{Mercedes B 170} \code{Mercedes B 200 CDI Autotronic (DPF)} \code{Mercedes B 200 Turbo Autotronic} \code{Mercedes C 180 Kompressor Elegance} \code{Mercedes C 180 Kompressor T-Modell Elegance} \code{Mercedes C 200 CDI T-Modell Classic} \code{Mercedes C 200 Kompressor Classic} \code{Mercedes C 200 Kompressor T-Modell Avantgarde Automatik} \code{Mercedes C 220 CDI Elegance (DPF)} \code{Mercedes C 220 CDI T-Modell Avantgarde (DPF)} \code{Mercedes C 220 CDI T-Modell Elegance (DPF)} \code{Mercedes C 230 T-Modell Avantgarde} \code{Mercedes C 280 Avantgarde} \code{Mercedes C 320 CDI Elegance 7G-Tronic (DPF)} \code{Mercedes CL 500 7G-Tronic} \code{Mercedes CL 500 Automatik} \code{Mercedes CLC 220 CDI (DPF)} \code{Mercedes CLK 220 CDI Coup<e9> Elegance (DPF)} \code{Mercedes CLK 270 CDI Coup<e9> Elegance Automatik} \code{Mercedes CLS 320 CDI 7G-Tronic (DPF)} \code{Mercedes CLS 63 AMG 7G-Tronic} \code{Mercedes E 200 Kompressor Classic} \code{Mercedes E 200 Kompressor T-Modell Elegance} \code{Mercedes E 220 CDI Classic (DPF)} \code{Mercedes E 220 CDI Elegance Automatik} \code{Mercedes E 220 CDI Elegance Automatik (DPF)} \code{Mercedes E 220 CDI T-Modell Elegance} \code{Mercedes E 280 CDI Elegance (DPF)} \code{Mercedes E 300 Bluetec Elegance 7G-Tronic} \code{Mercedes E 320 CDI Avantgarde Automatik} \code{Mercedes E 320 CDI Elegance 7G-Tronic (DPF)} \code{Mercedes E 320 CDI Elegance Automatik (DPF)} \code{Mercedes E 350 CGI Elegance 7G-Tronic} \code{Mercedes GL 320 CDI 4Matic (7G-Tronic) (DPF)} \code{Mercedes ML 270 CDI Automatik} \code{Mercedes S 320 CDI 7G-Tronic (DPF)} \code{Mercedes S 320 CDI Automatik (DPF)} \code{Mercedes S 350 7G-Tronic} \code{Mercedes S 350 Automatik} \code{Mercedes S 420 CDI 7G-Tronic (DPF)} \code{Mercedes SL 350 Sportmotor 7G-Tronic} \code{Mercedes SL 500 7G-Tronic} \code{Mercedes SLK 200 Kompressor} \code{Mercedes SLK 280} \code{Mercedes SLK 350} \code{Mercedes SLK 350 7G-Tronic} \code{Mercedes Sprinter Kombi 315 CDI kurz 3.5t Automatik (DPF)} \code{Mercedes Viano 2.0 CDI Trend kompakt (DPF)} \code{Mercedes Viano 2.2 CDI Fun kompakt} \code{MG TF 135} \code{Mini Cooper} \code{Mini Cooper Cabrio} \code{Mini Cooper D (DPF)} \code{Mini Cooper D Clubman (DPF)} \code{Mini Cooper S} \code{Mini Cooper S Clubman} \code{Mini One Cabrio} \code{Mini One Seven} \code{Mitsubishi Colt 1.5 DI-D Invite} \code{Mitsubishi Colt 1.5 DI-D Invite Automatik} \code{Mitsubishi Colt 1.5 Instyle} \code{Mitsubishi Colt CZ3 1.3 Invite} \code{Mitsubishi Grandis 2.0 DI-D Intense} \code{Mitsubishi Grandis 2.4 Intense} \code{Mitsubishi L200 2.5 DI-D Double Cab Intense Super Select 4WD} \code{Mitsubishi Lancer 1.6 Comfort} \code{Mitsubishi Lancer 1.8 MPI Invite} \code{Mitsubishi Lancer 2.0 DI-D Instyle (DPF)} \code{Mitsubishi Lancer Kombi 2.0 Sport} \code{Mitsubishi Outlander 2.0 DI-D Instyle (DPF offen)} \code{Mitsubishi Outlander 2.2 DI-D Instyle (DPF)} \code{Mitsubishi Pajero 3.2 DI-D Intense Automatik (DPF)} \code{Mitsubishi Pajero Classic 2.5 TD} \code{Nissan 350Z Coup<e9> Premium Pack} \code{Nissan 350Z Roadster Premium Pack} \code{Nissan Almera 1.5 acenta} \code{Nissan Almera Tino 1.8 acenta plus} \code{Nissan Almera Tino 2.2 dCi tekna} \code{Nissan Micra 1.2 Acenta} \code{Nissan Micra 1.2 visia} \code{Nissan Micra 1.2 Visia} \code{Nissan Micra 1.4 Acenta Sport} \code{Nissan Micra 1.5 dCi Tekna} \code{Nissan Micra C+C 1.6 Premium} \code{Nissan Navara Double Cab 2.5 dCi LE (DPF)} \code{Nissan Primera Traveller 1.9 dCi acenta} \code{Nissan Primera Traveller 2.0 tekna CVT-Automatik} \code{Nissan Qashqai 1.5 dCi DPF acenta 4x2} \code{Nissan Qashqai 1.6 acenta 4x2} \code{Nissan Qashqai 2.0 tekna 4x4} \code{Nissan Tiida 1.6 tekna} \code{Nissan Tiida 1.8 tekna} \code{Nissan X-Trail 2.0 dCi LE 4x4 (DPF)} \code{Nissan X-Trail 2.5 Elegance 4x4} \code{Nissan X-Trail 2.5 LE 4x4 CVT-Automatik} \code{Opel Agila 1.2 Edition} \code{Opel Agila 1.3 CDTI Enjoy} \code{Opel Antara 2.0 CDTI Edition (DPF)} \code{Opel Astra 1.3 CDTI (DPF)} \code{Opel Astra 1.4 Twinport Enjoy} \code{Opel Astra 1.6} \code{Opel Astra 1.6 Twinport Enjoy} \code{Opel Astra 1.7 CDTI Edition (6-Gang)} \code{Opel Astra 1.8 Sport} \code{Opel Astra Caravan 1.7 CDTI Cosmo (DPF)} \code{Opel Astra Caravan 1.9 CDTI Elegance} \code{Opel Astra GTC 1.6 Turbo Cosmo} \code{Opel Astra GTC 1.9 CDTI Cosmo (DPF)} \code{Opel Combo Combi 1.7 CDTI Sport} \code{Opel Combo Tour 1.3 CDTI} \code{Opel Corsa 1.0 Twinport Catch Me Now} \code{Opel Corsa 1.0 Twinport Edition} \code{Opel Corsa 1.2 Twinport Edition} \code{Opel Corsa 1.3 CDTI Cosmo} \code{Opel Corsa 1.7 DTI} \code{Opel GT 2.0 Turbo} \code{Opel Meriva 1.7 CDTI Cosmo} \code{Opel Meriva 1.8 Enjoy} \code{Opel Signum 2.2 direct Cosmo} \code{Opel Tigra TwinTop 1.3 CDTI Enjoy} \code{Opel Tigra TwinTop 1.4 Twinport Enjoy} \code{Opel Vectra 1.9 CDTI DPF Elegance} \code{Opel Vectra Caravan 1.9 CDTI DPF Edition} \code{Opel Vectra Caravan 2.2 direct Elegance} \code{Opel Vivaro Life 2.5 CDTI} \code{Opel Zafira 1.6 CNG Edition (Erdgasbetrieb)} \code{Opel Zafira 1.7 CDTI Edition (DPF)} \code{Opel Zafira 1.9 CDTI Cosmo (DPF)} \code{Opel Zafira 1.9 CDTI Edition Automatik (DPF)} \code{Opel Zafira 2.2 DTI Edition} \code{Peugeot 1007 110 Sport 2-Tronic} \code{Peugeot 1007 75 Filou} \code{Peugeot 1007 HDi FAP 110 Sport} \code{Peugeot 206 60 Grand Filou} \code{Peugeot 206 90 ESP Pr<e9>mium} \code{Peugeot 206 CC 135 Roland Garros} \code{Peugeot 206 CC HDi FAP 110 Tendance} \code{Peugeot 206 HDi <e9>co 70 Grand Filou} \code{Peugeot 207 110 Sport} \code{Peugeot 207 CC 150 THP Sport} \code{Peugeot 207 CC HDi FAP 110 Sport} \code{Peugeot 207 HDi FAP 110 Platinum} \code{Peugeot 207 RC Cup} \code{Peugeot 207 SW 75 Tendance} \code{Peugeot 307 110 Grand Filou Cool} \code{Peugeot 307 110 Tendance} \code{Peugeot 307 Break HDi 90 Esplanade} \code{Peugeot 307 Break HDi FAP 135 Tendance} \code{Peugeot 307 Break HDi FAP 90 Tendance} \code{Peugeot 307 CC 135} \code{Peugeot 307 CC HDi FAP 135 Sport} \code{Peugeot 307 HDi FAP 110 Platinum} \code{Peugeot 307 HDi FAP 135 Pr<e9>mium} \code{Peugeot 307 SW HDi FAP 110 Pr<e9>mium} \code{Peugeot 308 120 VTi Sport} \code{Peugeot 308 HDi FAP 110 Sport} \code{Peugeot 308 SW 120 VTi Sport Plus} \code{Peugeot 308 SW HDi FAP 135 Sport Plus} \code{Peugeot 4007 HDi FAP 155 Platinum (7-Sitzer)} \code{Peugeot 406 Break HDi FAP 135 Pr<e9>mium} \code{Peugeot 406 Break HPi 140 Sport} \code{Peugeot 406 Coup<e9> HDi FAP 135 Platinum} \code{Peugeot 407 115 Esplanade} \code{Peugeot 407 Coup<e9> V6 HDI FAP 205 Bi-Turbo Platinum Automatik} \code{Peugeot 407 HDi FAP 135 Tendance} \code{Peugeot 407 SW HDi FAP 135 Platinum} \code{Peugeot 407 SW V6 HDi FAP 205 Bi-Turbo Platinum Automatik} \code{Peugeot 607 V6 HDi FAP 205 Bi-Turbo Platinum Automatik} \code{Peugeot 807 HDi FAP 130 Tendance} \code{Peugeot 807 HDi FAP 135 Platinum} \code{Peugeot Partner Tepee HDi FAP 110 Escapade} \code{Porsche Boxster S 3.4} \code{Porsche Cayman S 3.4} \code{Renault Clio 1.2 16V Authentique} \code{Renault Clio 1.2 16V Confort Authentique} \code{Renault Clio 1.2 16V Privil<e8>ge Quickshift} \code{Renault Clio 1.5 dCi Confort Dynamique} \code{Renault Clio 1.5 dCi Dynamique} \code{Renault Clio 1.5 dCi ESP Edition Dynamique} \code{Renault Clio 1.5 dCi Initiale} \code{Renault Clio 1.6 16V Edition Dynamique} \code{Renault Espace 2.0 dCi FAP Privil<e8>ge} \code{Renault Grand Espace 3.0 dCi Initiale Automatik} \code{Renault Grand Modus 1.5 dCi ESP FAP Dynamique} \code{Renault Grand Sc<e9>nic 2.0 16V Emotion} \code{Renault Kangoo 1.6 16V Edition Campus} \code{Renault Kangoo 1.6 16V Privil<e8>ge} \code{Renault Laguna 2.0 16V Expression} \code{Renault Laguna 2.0 16V IDE Privil<e8>ge} \code{Renault Laguna 2.0 dCi FAP Dynamique} \code{Renault Laguna Grandtour 2.0 16V Dynamique} \code{Renault Laguna Grandtour 2.0 dCi FAP GT} \code{Renault Laguna Grandtour 2.0 dCi FAP Initiale} \code{Renault M<e9>gane 1.5 dCi FAP Dynamique} \code{Renault M<e9>gane 1.5 dCi Luxe Privil<e9>ge} \code{Renault M<e9>gane 1.6 16V Luxe Privil<e9>ge} \code{Renault M<e9>gane 1.9 dCi FAP Dynamique} \code{Renault M<e9>gane 2.0 16V Confort Privil<e9>ge} \code{Renault Modus 1.2 16V Avantage} \code{Renault Modus 1.6 16V Dynamique} \code{Renault Sc<e9>nic 1.6 16V Dynamique} \code{Renault Sc<e9>nic 1.6 16V Luxe Privil<e8>ge} \code{Renault Sc<e9>nic 1.9 dCi Luxe Dynamique} \code{Renault Trafic Passenger 2.0 dCi 115 Privil<e8>ge} \code{Renault Twingo 1.2 16V Dynamique} \code{Renault Twingo 1.2 16V TCE GT} \code{Renault Twingo 1.2 Authentique} \code{Renault Twingo 1.5 dCi Dynamique} \code{Renault Vel Satis 3.0 dCi Privil<e8>ge Automatik} \code{Renault Vel Satis 3.5 V6 Initiale Automatik} \code{Rover 75 2.0 CDTi Celeste S} \code{Saab 9-3 1.8i Vector} \code{Saab 9-3 Cabriolet 1.9 TTiD Aero (DPF)} \code{Saab 9-3 Cabriolet 2.0t Vector} \code{Saab 9-3 SportCombi 1.9 TiD Vector (DPF)} \code{Saab 9-3 SportCombi 2.0t Biopower Vector (Ethanol-Betrieb)} \code{Saab 9-3 SportCombi 2.0t Vector} \code{Saab 9-5 SportCombi 1.9 TiD Vector (DPF)} \code{Saab 9-5 SportCombi 2.0t Biopower Vector (Ethanol-Betrieb)} \code{Seat Altea 1.6 Reference} \code{Seat Altea 1.9 TDI Stylance} \code{Seat Altea XL 1.8T FSI Stylance} \code{Seat Altea XL 2.0 TDI Stylance (DPF)} \code{Seat Ibiza 1.2 12V Signo} \code{Seat Ibiza 1.4 TDI Signo} \code{Seat Ibiza 1.8 20V T Cupra} \code{Seat Le<f3>n 2.0 TDI Stylance (DPF)} \code{Seat Le<f3>n Cupra R} \code{Seat Toledo 1.9 TDI Signo} \code{Seat Toledo 2.0 FSI Stylance} \code{Seat Toledo 2.0 TDI Sport Edition DSG} \code{Skoda Fabia 1.2 HTP 12V Comfort} \code{Skoda Fabia 1.4 16V Elegance} \code{Skoda Fabia 1.4 TDI Elegance (DPF)} \code{Skoda Fabia 1.9 TDI RS} \code{Skoda Fabia Combi 1.4 TDI Ambiente (DPF)} \code{Skoda Fabia Combi 1.9 TDI Extra} \code{Skoda Octavia 1.6 FSI Ambiente} \code{Skoda Octavia 1.9 TDI Ambiente} \code{Skoda Octavia Combi 1.8 Turbo LK} \code{Skoda Octavia Combi 1.9 TDI Elegance} \code{Skoda Octavia Combi 2.0 FSI Elegance} \code{Skoda Octavia Combi 2.0 TDI Elegance DSG (DPF)} \code{Skoda Octavia Combi RS TDI (DPF)} \code{Skoda Roomster Scout 1.6 16V} \code{Skoda Superb 1.9 TDI Elegance} \code{Skoda Superb 2.0 TDI LK (DPF)} \code{Skoda Superb 2.8 V6 Elegance Tiptronic} \code{smart forfour 1.5 cdi pulse} \code{smart forfour 1.5 passion softouch plus} \code{smart forfour Brabus} \code{smart fortwo cabrio 0.8 cdi passion (DPF offen)} \code{smart fortwo coup<e9> 0.8 cdi pure (DPF offen)} \code{smart fortwo coup<e9> 1.0 mhd pulse} \code{smart fortwo coup<e9> 1.0 passion} \code{smart fortwo coup<e9> 1.0 pure} \code{smart fortwo coup<e9> 1.0 turbo passion} \code{smart fortwo coup<e9> passion} \code{SsangYong Actyon Sports 200 Xdi 4x4} \code{Subaru B9 Tribeca 3.0R Exclusive Automatik} \code{Subaru Forester 2.0X Exclusive} \code{Subaru Impreza 2.0R Comfort} \code{Subaru Justy 1.0 Active} \code{Subaru Legacy 3.0R Comfort Navigation Automatik} \code{Subaru Legacy Kombi 2.0D Comfort (DPF)} \code{Subaru Outback 2.5 ecomatic Comfort Automatik (Autogasbetrieb) (LPG)} \code{Subaru Outback 2.5 ecomatic Comfort Automatik (Benzinbetrieb)} \code{Suzuki Grand Vitara 1.9 DDiS Comfort+ (DPF)} \code{Suzuki Ignis 1.3 DDiS Comfort} \code{Suzuki Ignis 1.5 Comfort} \code{Suzuki Jimny 1.5 DDiS Comfort} \code{Suzuki Jimny Cabrio 1.3 Classic} \code{Suzuki Liana 1.4 DDiS Comfort} \code{Suzuki Liana 1.6 Comfort} \code{Suzuki Splash 1.0 Comfort} \code{Suzuki Splash 1.2 Comfort} \code{Suzuki Swift 1300 Comfort} \code{Suzuki Swift 1300 DDiS Comfort (DPF)} \code{Suzuki Swift 1600 Sport} \code{Suzuki Wagon R+ 1.3 DDiS Comfort} \code{Toyota Auris 1.6 Executive} \code{Toyota Auris 2.0 D-4D Executive (DPF)} \code{Toyota Auris 2.2 D-CAT} \code{Toyota Avensis 2.0 D-CAT Executive} \code{Toyota Avensis 2.0 Executive} \code{Toyota Avensis Combi 2.0 D-4D Executive} \code{Toyota Avensis Combi 2.2 D-CAT Executive} \code{Toyota Avensis Verso 2.0 Executive} \code{Toyota Aygo 1.0 City MMT} \code{Toyota Aygo 1.0 Club MMT} \code{Toyota Celica 1.8 TS} \code{Toyota Corolla 1.4 D-4D Sol} \code{Toyota Corolla 1.6 Executive} \code{Toyota Corolla 1.6 Sol Automatik} \code{Toyota Corolla 2.0 D-4D Sol} \code{Toyota Corolla Combi 1.6 Sol} \code{Toyota Corolla Verso 1.8 Executive} \code{Toyota Corolla Verso 1.8 Executive (7-Sitzer)} \code{Toyota Corolla Verso 1.8 linea sol} \code{Toyota Corolla Verso 2.2 D-4D Executive (7-Sitzer) (DPF)} \code{Toyota Corolla Verso 2.2 D-CAT Executive} \code{Toyota Hilux 3.0 D-4D Double Cab Exexcutive 4x4} \code{Toyota Land Cruiser 3.0 D-4D Executive Automatik} \code{Toyota MR2 Roadster SMT} \code{Toyota Previa 2.0 D-4D Sol} \code{Toyota Prius 1.5 Hybrid Executive} \code{Toyota Prius 1.5 Hybrid Sol} \code{Toyota RAV4 2.0 Sol 4x4} \code{Toyota RAV4 2.2 D-CAT Executive 4x4} \code{Toyota Yaris 1.0 Sol MMT} \code{Toyota Yaris 1.4 D-4D Executive} \code{Toyota Yaris 1.4 D-4D linea sol} \code{Toyota Yaris TS} \code{Volvo C30 1.6D Momentum (DPF)} \code{Volvo C30 1.8F Summum (Ethanol-Betrieb)} \code{Volvo C30 D5 Summum Geartronic (DPF)} \code{Volvo S40 2.0D Summum (DPF)} \code{Volvo S40 2.4i Momentum} \code{Volvo S60 D5 Comfort} \code{Volvo V50 1.6D Momentum (DPF)} \code{Volvo V50 2.0D Momentum} \code{Volvo V50 2.0D Summum (DPF)} \code{Volvo V70 2.0D Summum (DPF)} \code{Volvo V70 2.4 Bi-Fuel CNG Momentum (Benzinbetrieb)} \code{Volvo V70 2.4 Bi-Fuel CNG Momentum (Erdgasbetrieb)} \code{Volvo V70 T5 Summum} \code{Volvo XC90 D5 Executive Geartronic (7-Sitzer) (DPF)} \code{VW Caddy Life 1.6} \code{VW Caddy Life 1.9 TDI} \code{VW Caddy Life 1.9 TDI (DPF)} \code{VW CrossGolf 1.4 TSI} \code{VW CrossTouran 1.4 TSI} \code{VW Eos 2.0 TDI (DPF)} \code{VW Fox 1.2} \code{VW Fox 1.4 TDI} \code{VW Golf 1.4} \code{VW Golf 1.4 16V Trendline} \code{VW Golf 1.4 TSI Comfortline} \code{VW Golf 1.4 TSI Comfortline DSG (7-Gang)} \code{VW Golf 1.4 TSI GT Sport} \code{VW Golf 1.6 FSI Comfortline} \code{VW Golf 1.9 TDI Comfortline} \code{VW Golf 1.9 TDI Comfortline (DPF)} \code{VW Golf 1.9 TDI Sportline 4Motion} \code{VW Golf 2.0 FSI Comfortline} \code{VW Golf BlueMotion Trendline (DPF)} \code{VW Golf GTI DSG} \code{VW Golf Plus 1.4 TSI Comfortline} \code{VW Golf Plus 1.9 TDI Sportline} \code{VW Golf Variant 1.4 TSI Comfortline} \code{VW Jetta 2.0 TDI Comfortline} \code{VW Lupo 1.4} \code{VW New Beetle 1.9 TDI} \code{VW Passat 2.0 FSI Sportline} \code{VW Passat 2.0 TDI Highline DSG (DPF)} \code{VW Passat 2.0 TDI Sportline (DPF)} \code{VW Passat BlueMotion (DPF)} \code{VW Passat CC 1.8 TSI} \code{VW Passat Variant 2.0 5V Trendline} \code{VW Passat Variant 2.0 TDI Highline DSG (DPF)} \code{VW Passat Variant 2.0 TDI Trendline (DPF)} \code{VW Passat Variant 2.0 Turbo FSI Highline Tiptronic} \code{VW Passat Variant BlueMotion (DPF)} \code{VW Phaeton V6 Tiptronic (5-Sitzer)} \code{VW Polo 1.2} \code{VW Polo 1.4 FSI Highline} \code{VW Polo 1.4 TDI Trendline (DPF)} \code{VW Polo 1.6 Trendline} \code{VW Polo 1.9 TDI Sportline} \code{VW Polo BlueMotion (DPF)} \code{VW Polo Fun 1.4 TDI} \code{VW Polo Limousine 1.4 Comfortline} \code{VW Sharan 2.0 TDI Trendline (DPF)} \code{VW Sharan 2.8 V6 Highline Tiptronic} \code{VW T5 Caravelle 1.9 TDI Trendline kurz (DPF)} \code{VW Tiguan 2.0 TDI Sport Style (DPF)} \code{VW Touareg V6 FSI Tiptronic} \code{VW Touareg V6 TDI (DPF)} \code{VW Touareg V8 Tiptronic} \code{VW Touran 1.4 TSI Highline DSG} \code{VW Touran 1.9 TDI Trendline} \code{VW Touran 1.9 TDI Trendline DSG (DPF)} \code{VW Touran 2.0 FSI Trendline} \code{VW Touran EcoFuel Trendline (Erdgasbetrieb)}}
    \item{\code{V4}}{a factor with levels \code{aktuell} \code{Kleinstwagen} \code{Kleinwagen} \code{Microwagen} \code{Mittelklasse} \code{Obere Mittelklasse} \code{Oberklasse} \code{Untere Mittelklasse}}
    \item{\code{V5}}{a factor with levels \code{Ja} \code{Nein} \code{Steuerklasse}}
    \item{\code{V6}}{a factor with levels \code{Euro3} \code{Euro4} \code{Motor}}
    \item{\code{V7}}{a factor with levels \code{Diesel} \code{Gas} \code{Hybrid} \code{kW} \code{Otto} \code{Wankel}}
    \item{\code{V8}}{a factor with levels \code{100} \code{103} \code{104} \code{105} \code{106} \code{107} \code{108} \code{110} \code{112} \code{114} \code{115} \code{118} \code{119} \code{120} \code{121} \code{122} \code{124} \code{125} \code{126} \code{127} \code{128} \code{129} \code{130} \code{131} \code{132} \code{133} \code{135} \code{136} \code{140} \code{141} \code{142} \code{143} \code{145} \code{147} \code{149} \code{150} \code{153} \code{155} \code{160} \code{162} \code{165} \code{169} \code{170} \code{171} \code{173} \code{175} \code{176} \code{177} \code{180} \code{190} \code{191} \code{194} \code{195} \code{196} \code{200} \code{203} \code{206} \code{208} \code{210} \code{215} \code{217} \code{225} \code{228} \code{232} \code{235} \code{240} \code{242} \code{245} \code{250} \code{253} \code{255} \code{280} \code{285} \code{306} \code{309} \code{327} \code{33} \code{373} \code{378} \code{38} \code{40} \code{43} \code{44} \code{45} \code{46} \code{47} \code{48} \code{49} \code{50} \code{51} \code{52} \code{54} \code{55} \code{56} \code{59} \code{60} \code{61} \code{62} \code{63} \code{64} \code{65} \code{66} \code{67} \code{68} \code{69} \code{70} \code{71} \code{72} \code{73} \code{74} \code{75} \code{76} \code{77} \code{78} \code{79} \code{80} \code{81} \code{82} \code{83} \code{84} \code{85} \code{88} \code{89} \code{90} \code{91} \code{92} \code{93} \code{94} \code{95} \code{96} \code{98} \code{99} \code{CCM}}
    \item{\code{V9}}{a factor with levels \code{1086} \code{1108} \code{1120} \code{1124} \code{1149} \code{1198} \code{1206} \code{1229} \code{1240} \code{1242} \code{1248} \code{1298} \code{1308} \code{1328} \code{1332} \code{1339} \code{1348} \code{1349} \code{1360} \code{1364} \code{1368} \code{1386} \code{1388} \code{1390} \code{1396} \code{1398} \code{1399} \code{1422} \code{1461} \code{1468} \code{1490} \code{1493} \code{1495} \code{1497} \code{1498} \code{1499} \code{1560} \code{1582} \code{1584} \code{1586} \code{1587} \code{1590} \code{1591} \code{1595} \code{1596} \code{1598} \code{1599} \code{1686} \code{1689} \code{1699} \code{1749} \code{1753} \code{1769} \code{1781} \code{1793} \code{1794} \code{1796} \code{1798} \code{1799} \code{1840} \code{1870} \code{1896} \code{1910} \code{1951} \code{1968} \code{1970} \code{1975} \code{1984} \code{1985} \code{1991} \code{1992} \code{1994} \code{1995} \code{1997} \code{1998} \code{1999} \code{2034} \code{2148} \code{2171} \code{2172} \code{2179} \code{2184} \code{2188} \code{2198} \code{2204} \code{2231} \code{2261} \code{2351} \code{2354} \code{2359} \code{2378} \code{2387} \code{2393} \code{2399} \code{2400} \code{2401} \code{2429} \code{2435} \code{2457} \code{2463} \code{2477} \code{2488} \code{2492} \code{2494} \code{2496} \code{2497} \code{2499} \code{2500} \code{2521} \code{2597} \code{2685} \code{2698} \code{2720} \code{2736} \code{2771} \code{2777} \code{2792} \code{2902} \code{2958} \code{2967} \code{2976} \code{2979} \code{2982} \code{2987} \code{2993} \code{2995} \code{2996} \code{3000} \code{3189} \code{3195} \code{3197} \code{3199} \code{3200} \code{3222} \code{3311} \code{3342} \code{3387} \code{3456} \code{3471} \code{3497} \code{3498} \code{3597} \code{3724} \code{3778} \code{3996} \code{3999} \code{4134} \code{4163} \code{4172} \code{4196} \code{4293} \code{4398} \code{4423} \code{4608} \code{4966} \code{4969} \code{4999} \code{5461} \code{5654} \code{5666} \code{6208} \code{659} \code{698} \code{796} \code{799} \code{989} \code{995} \code{996} \code{998} \code{999} \code{Verbrauch}}
    \item{\code{V10}}{a factor with levels \code{10.02} \code{10.03} \code{10.05} \code{10.08} \code{10.09} \code{10.12} \code{10.14} \code{10.18} \code{10.19} \code{10.29} \code{10.33} \code{10.34} \code{10.41} \code{10.43} \code{10.47} \code{10.55} \code{10.58} \code{10.6} \code{10.62} \code{10.63} \code{10.68} \code{10.69} \code{10.7} \code{10.78} \code{10.8} \code{10.84} \code{10.89} \code{10.94} \code{10.98} \code{11.13} \code{11.22} \code{11.23} \code{11.24} \code{11.28} \code{11.32} \code{11.33} \code{11.37} \code{11.39} \code{11.4} \code{11.71} \code{11.9} \code{11.96} \code{12.09} \code{12.14} \code{12.2} \code{12.21} \code{12.22} \code{12.3} \code{12.36} \code{12.4} \code{12.44} \code{12.49} \code{12.86} \code{13.43} \code{13.85} \code{14.04} \code{14.13} \code{14.54} \code{4.03} \code{4.38} \code{4.41} \code{4.5} \code{4.54} \code{4.63} \code{4.7} \code{4.71} \code{4.75} \code{4.78} \code{4.81} \code{4.82} \code{4.83} \code{4.86} \code{4.89} \code{4.93} \code{4.95} \code{4.96} \code{4.97} \code{4.98} \code{4.99} \code{5.02} \code{5.04} \code{5.06} \code{5.07} \code{5.1} \code{5.15} \code{5.17} \code{5.18} \code{5.19} \code{5.2} \code{5.21} \code{5.22} \code{5.23} \code{5.25} \code{5.26} \code{5.27} \code{5.28} \code{5.29} \code{5.31} \code{5.32} \code{5.33} \code{5.38} \code{5.43} \code{5.44} \code{5.45} \code{5.46} \code{5.48} \code{5.49} \code{5.5} \code{5.51} \code{5.53} \code{5.54} \code{5.56} \code{5.57} \code{5.58} \code{5.59} \code{5.6} \code{5.62} \code{5.64} \code{5.65} \code{5.66} \code{5.67} \code{5.68} \code{5.69} \code{5.7} \code{5.71} \code{5.73} \code{5.74} \code{5.75} \code{5.76} \code{5.77} \code{5.78} \code{5.79} \code{5.8} \code{5.81} \code{5.82} \code{5.84} \code{5.85} \code{5.86} \code{5.87} \code{5.88} \code{5.9} \code{5.91} \code{5.92} \code{5.93} \code{5.94} \code{5.95} \code{5.96} \code{5.97} \code{5.98} \code{5.99} \code{6.0} \code{6.02} \code{6.04} \code{6.05} \code{6.06} \code{6.07} \code{6.08} \code{6.09} \code{6.1} \code{6.12} \code{6.13} \code{6.14} \code{6.15} \code{6.16} \code{6.18} \code{6.19} \code{6.2} \code{6.21} \code{6.22} \code{6.23} \code{6.24} \code{6.26} \code{6.27} \code{6.28} \code{6.29} \code{6.3} \code{6.31} \code{6.32} \code{6.34} \code{6.35} \code{6.36} \code{6.38} \code{6.39} \code{6.4} \code{6.41} \code{6.42} \code{6.44} \code{6.45} \code{6.46} \code{6.49} \code{6.5} \code{6.52} \code{6.53} \code{6.54} \code{6.55} \code{6.56} \code{6.58} \code{6.59} \code{6.6} \code{6.61} \code{6.62} \code{6.63} \code{6.64} \code{6.65} \code{6.67} \code{6.68} \code{6.69} \code{6.7} \code{6.71} \code{6.72} \code{6.73} \code{6.74} \code{6.75} \code{6.76} \code{6.77} \code{6.78} \code{6.79} \code{6.8} \code{6.81} \code{6.82} \code{6.83} \code{6.84} \code{6.85} \code{6.86} \code{6.87} \code{6.89} \code{6.9} \code{6.91} \code{6.92} \code{6.94} \code{6.95} \code{6.96} \code{6.97} \code{6.98} \code{6.99} \code{7.0} \code{7.01} \code{7.03} \code{7.04} \code{7.05} \code{7.06} \code{7.07} \code{7.08} \code{7.09} \code{7.1} \code{7.11} \code{7.14} \code{7.15} \code{7.16} \code{7.19} \code{7.2} \code{7.21} \code{7.23} \code{7.24} \code{7.25} \code{7.27} \code{7.29} \code{7.3} \code{7.31} \code{7.32} \code{7.33} \code{7.34} \code{7.35} \code{7.36} \code{7.38} \code{7.39} \code{7.41} \code{7.42} \code{7.43} \code{7.44} \code{7.46} \code{7.47} \code{7.48} \code{7.49} \code{7.5} \code{7.51} \code{7.52} \code{7.53} \code{7.54} \code{7.55} \code{7.56} \code{7.57} \code{7.58} \code{7.59} \code{7.6} \code{7.61} \code{7.63} \code{7.64} \code{7.66} \code{7.68} \code{7.69} \code{7.71} \code{7.73} \code{7.74} \code{7.75} \code{7.76} \code{7.77} \code{7.78} \code{7.8} \code{7.82} \code{7.83} \code{7.84} \code{7.85} \code{7.89} \code{7.9} \code{7.91} \code{7.92} \code{7.93} \code{7.96} \code{7.98} \code{8.0} \code{8.01} \code{8.02} \code{8.03} \code{8.04} \code{8.07} \code{8.09} \code{8.1} \code{8.11} \code{8.12} \code{8.13} \code{8.14} \code{8.15} \code{8.16} \code{8.18} \code{8.19} \code{8.2} \code{8.21} \code{8.24} \code{8.26} \code{8.27} \code{8.28} \code{8.31} \code{8.32} \code{8.33} \code{8.34} \code{8.35} \code{8.36} \code{8.37} \code{8.38} \code{8.39} \code{8.4} \code{8.43} \code{8.44} \code{8.45} \code{8.46} \code{8.47} \code{8.48} \code{8.49} \code{8.5} \code{8.51} \code{8.52} \code{8.53} \code{8.54} \code{8.57} \code{8.58} \code{8.61} \code{8.62} \code{8.63} \code{8.64} \code{8.65} \code{8.66} \code{8.67} \code{8.68} \code{8.69} \code{8.7} \code{8.71} \code{8.72} \code{8.73} \code{8.74} \code{8.76} \code{8.77} \code{8.78} \code{8.79} \code{8.8} \code{8.81} \code{8.82} \code{8.83} \code{8.85} \code{8.86} \code{8.88} \code{8.89} \code{8.91} \code{8.92} \code{8.93} \code{8.94} \code{8.95} \code{8.96} \code{8.97} \code{8.98} \code{9.0} \code{9.01} \code{9.02} \code{9.04} \code{9.06} \code{9.07} \code{9.08} \code{9.1} \code{9.12} \code{9.13} \code{9.15} \code{9.16} \code{9.17} \code{9.21} \code{9.26} \code{9.29} \code{9.3} \code{9.31} \code{9.32} \code{9.35} \code{9.36} \code{9.4} \code{9.41} \code{9.42} \code{9.44} \code{9.45} \code{9.46} \code{9.47} \code{9.48} \code{9.52} \code{9.53} \code{9.54} \code{9.55} \code{9.59} \code{9.6} \code{9.64} \code{9.67} \code{9.68} \code{9.71} \code{9.83} \code{9.85} \code{9.86} \code{9.87} \code{9.92} \code{9.95} \code{9.96} \code{9.99} \code{Schadstoffe}}
    \item{\code{V11}}{a factor with levels \code{0} \code{11} \code{13} \code{14} \code{16} \code{17} \code{18} \code{19} \code{20} \code{21} \code{22} \code{23} \code{24} \code{25} \code{26} \code{27} \code{28} \code{29} \code{30} \code{31} \code{32} \code{33} \code{34} \code{35} \code{36} \code{37} \code{38} \code{39} \code{40} \code{41} \code{42} \code{43} \code{44} \code{45} \code{46} \code{47} \code{48} \code{49} \code{50} \code{CO2}}
    \item{\code{V12}}{a factor with levels \code{0} \code{1} \code{10} \code{11} \code{12} \code{13} \code{14} \code{15} \code{16} \code{17} \code{18} \code{19} \code{2} \code{20} \code{21} \code{22} \code{23} \code{24} \code{25} \code{26} \code{27} \code{28} \code{29} \code{3} \code{30} \code{31} \code{32} \code{33} \code{34} \code{35} \code{36} \code{37} \code{38} \code{39} \code{4} \code{40} \code{41} \code{43} \code{5} \code{6} \code{7} \code{8} \code{9} \code{gesamt}}
    \item{\code{V13}}{a factor with levels \code{19} \code{22} \code{25} \code{27} \code{29} \code{30} \code{32} \code{34} \code{35} \code{36} \code{37} \code{38} \code{39} \code{40} \code{41} \code{44} \code{45} \code{46} \code{47} \code{48} \code{49} \code{50} \code{51} \code{52} \code{53} \code{54} \code{55} \code{56} \code{57} \code{58} \code{59} \code{60} \code{61} \code{62} \code{63} \code{64} \code{65} \code{66} \code{67} \code{68} \code{69} \code{70} \code{71} \code{72} \code{73} \code{74} \code{75} \code{76} \code{77} \code{78} \code{79} \code{80} \code{81} \code{82} \code{83} \code{87} \code{89} \code{km}}
    \item{\code{V14}}{a factor with levels \code{1} \code{10} \code{11} \code{12} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{mc}}
    \item{\code{V15}}{a factor with levels \code{1} \code{10} \code{11} \code{12} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{hc1}}
    \item{\code{V16}}{a factor with levels \code{1} \code{10} \code{11} \code{12} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{hc2}}
    \item{\code{V17}}{a factor with levels \code{1} \code{10} \code{11} \code{12} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{kmO}}
    \item{\code{V18}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E} \code{F} \code{G} \code{H} \code{I} \code{J} \code{K} \code{L} \code{mcO}}
    \item{\code{V19}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E} \code{F} \code{G} \code{H} \code{hc1O} \code{I} \code{J} \code{K} \code{L}}
    \item{\code{V20}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E} \code{F} \code{G} \code{H} \code{hc2O} \code{I} \code{J} \code{K} \code{L}}
    \item{\code{V21}}{a factor with levels \code{} \code{A} \code{B} \code{C} \code{D} \code{E} \code{F} \code{G} \code{H} \code{I} \code{J} \code{K} \code{L}}
  }
}
\source{
www.adac.de
}
\examples{
data(eco2plus)
## maybe str(eco2plus) ; plot(eco2plus) ...
}
\keyword{datasets}
