\name{extracat}
\alias{extracat}

\docType{package}
\title{
Graphical extensions for catagorical data
}
\description{
Provides interactive parallel coordinates plots based on package \code{iplots} for categorical and continuous data
as well as a grid-based special version of multiple barcharts, \link[extracat]{rmb} which shows the relative frequencies of a target variable an the corresponding weights in a MB-like graphical layout.
Functions related to the Google Summer of Code 2011 are also included:
\link[extracat]{optile} resorts the categories of variables in order to sort of diagonalize the tables and graphical representations, which makes them easier to interpret. \link[extracat]{fluctile} provides fluctuation diagrams for the visualisation.
}
\details{
\tabular{ll}{
Package: \tab extracat\cr
Type: \tab Package\cr
Version: \tab 1.5-0\cr
Date: \tab 2012-04-25\cr
License: \tab -\cr
LazyLoad: \tab yes\cr
}

}
\author{
Alexander Pilhoefer\cr
Department for Computer Oriented Statistics and Data Analysis\cr
University of Augsburg\cr
Germany

Maintainer: Alexander Pilhoefer \email{alexander.pilhoefer@math.uni-augsburg.de}
}
\references{
Alexander Pilhoefer  \emph{Analysen von kategoriellen Daten}\cr Diploma Thesis 2009\cr
\cr
  Alexander Pilhoefer \emph{ New approaches in visualization of categorical data:
       R-package extracat}\cr
  Journal of Statistical Software, accepted 2011

}

\keyword{ categorical }
\keyword{ parallel coordinates }
\keyword{ multiple barcharts }
\keyword{ mosaicplots }
\keyword{ interactive }

