\name{GeneEx}
\alias{GeneEx}
\docType{data}
\title{
Gene Expression Data
}
\description{
Gene expression dataset.
}
\usage{data(GeneEx)}
\format{
  A data frame with 7705 observations (genes) on the following 52 variables (samples).
  \describe{
    \item{\code{CLID}}{clinical ID}
        \item{\code{NAME}}{a numeric vector}
    \item{\code{shee177}}{a numeric vector}
    \item{\code{shfa047}}{a numeric vector}
    \item{\code{shfs151}}{a numeric vector}
    \item{\code{shfa044}}{a numeric vector}
    \item{\code{shee146}}{a numeric vector}
    \item{\code{shee129}}{a numeric vector}
    \item{\code{shee118}}{a numeric vector}
    \item{\code{shee109}}{a numeric vector}
    \item{\code{shee100}}{a numeric vector}
    \item{\code{shet058}}{a numeric vector}
    \item{\code{shet057}}{a numeric vector}
    \item{\code{shet033}}{a numeric vector}
    \item{\code{shee045}}{a numeric vector}
    \item{\code{shco045}}{a numeric vector}
    \item{\code{shco044}}{a numeric vector}
    \item{\code{shco039}}{a numeric vector}
    \item{\code{shco038}}{a numeric vector}
    \item{\code{shco031}}{a numeric vector}
    \item{\code{shco030}}{a numeric vector}
    \item{\code{shco029}}{a numeric vector}
    \item{\code{shfe113}}{a numeric vector}
    \item{\code{shfe085}}{a numeric vector}
    \item{\code{shfe084}}{a numeric vector}
    \item{\code{shfe081}}{a numeric vector}
    \item{\code{shfe052}}{a numeric vector}
    \item{\code{shcz090}}{a numeric vector}
    \item{\code{shcz089}}{a numeric vector}
    \item{\code{shcz087}}{a numeric vector}
    \item{\code{shcz086}}{a numeric vector}
    \item{\code{shcz085}}{a numeric vector}
    \item{\code{shcz064}}{a numeric vector}
    \item{\code{shcz061}}{a numeric vector}
    \item{\code{shdj104}}{a numeric vector}
    \item{\code{shdj102}}{a numeric vector}
    \item{\code{shdj094}}{a numeric vector}
    \item{\code{shdj089}}{a numeric vector}
    \item{\code{shdj067}}{a numeric vector}
    \item{\code{shfs240}}{a numeric vector}
    \item{\code{shfs229}}{a numeric vector}
    \item{\code{shfs227}}{a numeric vector}
    \item{\code{shfs226}}{a numeric vector}
    \item{\code{shfa109}}{a numeric vector}
    \item{\code{shfa107}}{a numeric vector}
    \item{\code{shee216}}{a numeric vector}
    \item{\code{shet172}}{a numeric vector}
    \item{\code{shee198}}{a numeric vector}
    \item{\code{shfa067}}{a numeric vector}
    \item{\code{shfa066}}{a numeric vector}
    \item{\code{shee188}}{a numeric vector}
    \item{\code{shfs159}}{a numeric vector}
  }
}
\details{
See source.
}
\source{
\url{http://seurat.r-forge.r-project.org/}
}

\examples{
data(GeneEx)
## maybe str(GeneEx) ; plot(GeneEx) ...
}
\keyword{datasets}
