\name{MJnew}
\alias{MJnew}
\docType{data}
\title{
Example Matrix Data
}
\description{
US Convictions
}
\usage{data("MJnew")}
\format{
  A data frame with 508 observations on the following 107 variables.
  \describe{
    \item{\code{Last.Name}}{a factor with levels \code{Aatique} \code{Abassi} \code{Abdall} \code{Abdallah} \code{Abdhir} \code{Abdi} \code{Abdoulah} \code{Abdow} \code{Abdul-Latif} \code{Abdullahu} \code{Abdulmutallab} \code{Abdulqader} \code{Abdur-Raheem} \code{Abodayah} \code{Abraham} \code{Abshir} \code{Abu-Zbaida} \code{Abu Ali} \code{Abu Baker} \code{Abuali} \code{Abujihaad} \code{Afridi} \code{Afzali} \code{Ahissou} \code{Ahmad} \code{Ahmed} \code{Ahmedzay} \code{Akhdar} \code{Akl} \code{al-Aboody} \code{al-Aboudy} \code{Al-Amoudi} \code{al-Arian} \code{al-Atabi} \code{al-Attas} \code{al-Bakri} \code{al-Baraa} \code{al-Buthe} \code{al-Draibi} \code{al-Ghazi} \code{al-Hachami} \code{Al-Hadi} \code{al-Hamdan} \code{Al-Hamdi} \code{al-Kassar} \code{al-Khaledy} \code{al-Maleky} \code{al-Marabh} \code{al-Marri} \code{al-Moayad} \code{al-Nalfi} \code{al-Obaidi} \code{al-Qaryuti} \code{al-Rababah} \code{al-Salmi} \code{al-Saraf} \code{al-Shiblawy} \code{al-Tamimi} \code{al-Timimi} \code{Al-Uqaily} \code{al-Wahaidy} \code{Al Delaema} \code{Al Khammash} \code{Alandon} \code{Alazawi} \code{Albanna} \code{Albred} \code{Aldawsari} \code{Alessa} \code{Alferahin} \code{Alghamdi} \code{Ali} \code{Alibrahimi} \code{Alishtari} \code{Almari} \code{Almazaal} \code{Almonte} \code{Almosaleh} \code{Almutairi} \code{Alshomary} \code{Alubeidy} \code{Alvarez-Duenas} \code{Alwan} \code{Amawi} \code{Aref} \code{Arnaout} \code{Ashqar} \code{Asllani} \code{Atta} \code{Augustin} \code{Augustine} \code{Awad} \code{Awada} \code{Awan} \code{Awde} \code{Ayesh} \code{Azmath} \code{Babar} \code{Badri} \code{Bakarbashat} \code{Barrera-De Amaris} \code{Basnan} \code{Batiste} \code{Battle} \code{Bautista Martinez} \code{Benkahla} \code{Benmoumen} \code{Berjaoui} \code{Berro} \code{Biheiri} \code{Bilal} \code{Bin Osman} \code{Boughader-Mucharrafille} \code{Bouraima} \code{Bowman} \code{Boyd} \code{Brent} \code{Budiman} \code{Bujol} \code{Butt} \code{Carpenter} \code{Chandia} \code{Chapman} \code{Chesser} \code{Chhun} \code{Chughtai} \code{Cleveland} \code{Conde Rubio} \code{Cromitie} \code{Daher} \code{Damra} \code{Dato} \code{Dawe} \code{Defreitas} \code{Dhafir} \code{Dhakane} \code{Doudzai} \code{Duka} \code{Dweikat} \code{El-Hanafi} \code{el-Hindi} \code{el-Mezain} \code{el-Najjar} \code{El-Sablani} \code{el-Shukrijumah} \code{El-Shwehdi} \code{El Hage} \code{El Ouariachi} \code{El Zaatari} \code{Elahwal} \code{Elashi} \code{Elashmouny} \code{Elashyi} \code{Elfgeeh} \code{Elgindy} \code{Elmaghraby} \code{Elshafay} \code{Elzahabi} \code{Faarax} \code{Fall} \code{Farhane} \code{Farhat} \code{Farid} \code{Faris} \code{Fariz} \code{Fawaz} \code{Ferrari} \code{Fidse} \code{Finton} \code{Florez-Gomez} \code{Foley} \code{Ford} \code{Galab} \code{Galicia} \code{Gamarra-Murillo} \code{Gamboa} \code{Goba} \code{Godoy} \code{Gonzalez-Neira} \code{Gordon} \code{Grecula} \code{Grinberg} \code{Guagni} \code{Gujar} \code{Haddoumi} \code{Hafaiedh} \code{Hamdan} \code{Hameed} \code{Hammadi} \code{Hammami} \code{Hammoud} \code{Hansen} \code{Harb} \code{Hariri} \code{Hasan} \code{Hasanoff} \code{Hashmi} \code{Hassan} \code{Hassoun} \code{Hatfield} \code{Hawash} \code{Hayat} \code{Haydous} \code{Hazime} \code{Headley} \code{Hmimssa} \code{Hodroj} \code{Hossain} \code{Hupper} \code{Hussain} \code{Hussein} \code{Hyder} \code{Ibrahim} \code{Imam} \code{Iqbal} \code{Ismail} \code{Issa} \code{Isse} \code{Jabarah} \code{Jaber} \code{Jamal} \code{James} \code{Jarwan} \code{Jayyousi} \code{Jensen} \code{Johnson} \code{Jubara} \code{Kadir} \code{Kamran} \code{Kandasamy} \code{Kane} \code{Karaki} \code{Kassem} \code{Kassir} \code{Katz} \code{Kaziu} \code{Khafagi} \code{Khalil} \code{Khamis} \code{Khan} \code{Khatib} \code{Kilfat} \code{Kishk} \code{Kopke} \code{Koshak} \code{Kourani} \code{Kwon} \code{Laimeche} \code{Lakhani} \code{LaRose} \code{Lewis} \code{Lindh} \code{Londono} \code{Lopez} \code{Lopez-Flores} \code{Maatouk} \code{Maddy} \code{Maflahi} \code{Mahmood} \code{Mahmoud} \code{Makki} \code{Maldonado} \code{Malike} \code{Martinez} \code{Martinez-Flores} \code{Maruf} \code{Masfaka} \code{Masri} \code{Mazloum} \code{Medunjanin} \code{Mehanna} \code{Menepta} \code{Mohamed} \code{Mohammad} \code{Mohamud} \code{Moheisen} \code{Mora} \code{Morales} \code{Mosed} \code{Mostafa} \code{Moussaoui} \code{Mubayyid} \code{Mujahidh} \code{Muntasser} \code{Murshed} \code{Muthana} \code{Nabil} \code{Naeem} \code{Naseer} \code{Nasr} \code{Nasrallah} \code{Nasser} \code{Nguyen} \code{Nour} \code{Nur} \code{Odeh} \code{Omar} \code{Omian} \code{Ouazzani} \code{Oulai} \code{Padilla} \code{Paracha} \code{Patpanathan} \code{Patterson} \code{Paul} \code{Paulin-Ramirez} \code{Payen} \code{Pervez} \code{Peters} \code{Phanor} \code{Pierre} \code{Pineda} \code{Ponton Caro} \code{Pratheepan} \code{Puerta} \code{Qayyum} \code{Qazi} \code{Qureshi} \code{Rahal} \code{Rahmani} \code{Raja} \code{Ranjha} \code{Ranson} \code{Raza} \code{Refai} \code{Rehman} \code{Reid} \code{Reynolds} \code{Rizvi} \code{Rockwood} \code{Rodriguez} \code{Rodriguez-Acevedo} \code{Romero-Panchano} \code{Rouissi} \code{Royer} \code{Ruiz} \code{Rusli} \code{Saade} \code{Sabaratnam} \code{Sabir} \code{Sadequee} \code{Safadi} \code{Salah} \code{Salamanca} \code{Salat} \code{Sallam} \code{Samana} \code{Samhan} \code{Sandhu} \code{Sarachandran} \code{Sarama} \code{Sarsour} \code{Sarwer} \code{Sattar} \code{Sedaghaty} \code{Seif} \code{Serrano-Valdez} \code{Shah} \code{Shahzad} \code{Shaikh} \code{Shareef} \code{Shehadeh} \code{Sheikh} \code{Sherifi} \code{Shishani} \code{Shnewer} \code{Shorbagi} \code{Siddiqui} \code{Siraj} \code{Smadi} \code{Socrates} \code{Soedirdja} \code{Steitiye} \code{Stewart} \code{Subandi} \code{Subasic} \code{Subeh} \code{Sudani} \code{Suleiman} \code{Surratt} \code{Taher} \code{Talavera} \code{Taleb-Jedi} \code{Talyi} \code{Tapasco Romero} \code{Tatar} \code{Tauseef} \code{Tebbakh} \code{Thanigasalam} \code{Trabelsi} \code{Tumer} \code{Ujaama} \code{Ulloa Melo} \code{Upshur} \code{Varatharasa} \code{Varela} \code{Villalobos} \code{Vinas} \code{Vinayagamoorthy} \code{Wagner} \code{Walker} \code{Warsame} \code{Washington} \code{Wasil} \code{Williams} \code{Wingate} \code{Wotulo} \code{Wu} \code{Yaghi} \code{Yograrasa} \code{Younes} \code{Younis} \code{Yousry} \code{Yusuf} \code{Zayed} \code{Zazi} \code{Zeb}}
    \item{\code{First.Name.s.}}{a factor with levels \code{Aafia} \code{Abad} \code{Abdel Hameed} \code{Abdelhaleem} \code{Abdifatah Yusuf} \code{Abdikadir Ali} \code{Abdillah} \code{Abdisalan Hussein} \code{Abdiweli Yassin} \code{Abdow Munye} \code{Abdul} \code{Abdul Halim} \code{Abdul Karim Akram} \code{Abdul Rahman} \code{Abdul Tawala Ibn Ali} \code{Abdulamir} \code{Abdulghefur Abdul} \code{Abdulhakeem} \code{Abdulla Kasem} \code{Abdullahi Omar} \code{Abdulrahman} \code{Abdurahman} \code{Abid} \code{Abu Khalid} \code{Adarus Abdulle} \code{Adel} \code{Adham Admin} \code{Adis} \code{Adnan} \code{Adriana Gladys} \code{Agron} \code{Agus} \code{Ahmad} \code{Ahmad Abeed} \code{Ahmad Wais} \code{Ahmed} \code{Ahmed Abdel} \code{Ahmed Abdellatif Sherif} \code{Ahmed Abdulkadir} \code{Ahmed Ali} \code{Ahmed Hassan} \code{Ahmed Ibrahim} \code{Ahmed Muhammed} \code{Ahmed Nawaz} \code{Ahmed Omar} \code{Aisha} \code{Akeel} \code{Akram} \code{Akram Abdul Karim} \code{Alaa Allia Ahmed} \code{Alam} \code{Alawi Hussain} \code{Ali} \code{Ali Abdul Karim} \code{Ali Abdulamir} \code{Ali Asad} \code{Ali F.} \code{Ali Khaled} \code{Ali Mohamad} \code{Ali Mohammed Abboud} \code{Ali Saleh Kalah} \code{Almire Ali-Sadek} \code{Amera} \code{Amina} \code{Amina Farah} \code{Amira Ali} \code{Amna} \code{Amr I.} \code{Anes} \code{Ansar} \code{Antonio} \code{Aref} \code{Arkan} \code{Armoghan} \code{Arsalan Absar} \code{Arshad} \code{Arwah} \code{Ashar Iqbal} \code{Assam} \code{Atif} \code{Awais} \code{Ayman} \code{Bajram} \code{Barry} \code{Basman} \code{Bassam Kamel} \code{Bayan} \code{Ben Sami Fathi} \code{Bernardo Valdes} \code{Betim} \code{Bilal} \code{Brandy Jo} \code{Bryant Neal} \code{Burson} \code{Cabdulaahi Ahmed} \code{Carlos} \code{Carlos Ali Romero} \code{Carlos Eduardo} \code{Carmen Maria} \code{Carole} \code{Cedric} \code{Chao Tung} \code{Chawki} \code{Choudhry} \code{Christopher} \code{Colleen} \code{Corneille} \code{Daniel Joseph} \code{Daniel Patrick} \code{David} \code{David Coleman} \code{Deka Abdalla} \code{Derrick} \code{Dib Hani} \code{Donald Thomas} \code{Dritan} \code{Dylan} \code{Earnest James} \code{Edgar Fernando Blanco} \code{Edizon Ramirez} \code{Ehab} \code{Ehsanul Islam} \code{Elias Mohamad} \code{Eljvir} \code{Elkin Alberto Arroyave} \code{Elmeliani} \code{Emadeddin} \code{Emilio} \code{Enaam M.} \code{Erick} \code{Eyad} \code{Fadhil} \code{Fadi} \code{Fadl Mohammed} \code{Faisal} \code{Faisal M.} \code{Fanny} \code{Farida} \code{Farooque} \code{Fawaz} \code{Faycal} \code{Faysal} \code{Ferid} \code{Francis Sourou} \code{Francois} \code{G. William} \code{Ghassan} \code{Gregory} \code{Hadir} \code{Hafiz} \code{Hafiz Khalil} \code{Hafiz Muhammad Sher Ali} \code{Haider} \code{Haji} \code{Hamid} \code{Hammad} \code{Hammad Raiz} \code{Hamze} \code{Haniffa} \code{Hasan Ali} \code{Hasan Antar} \code{Hassan} \code{Hassan Moussa} \code{Hatef} \code{Hatem Naji} \code{Hawo Mohamed} \code{Hazim} \code{Hector} \code{Helmi} \code{Hemant} \code{Herbert} \code{Hisham} \code{Hor} \code{Hosam} \code{Hosam Yousef} \code{Hossain} \code{Houda Mohamad} \code{Husam Addelhafiz} \code{Hussain} \code{Hussein} \code{Hussein A.} \code{Hysen} \code{Ibrahim Ahmed} \code{Iftikhar} \code{Ihsan} \code{Ilyas} \code{Imdad Ullah} \code{Imran} \code{Imtiaz} \code{Irfan} \code{Issam Abdul-Hakim} \code{Issam Hassan} \code{Iyman} \code{Izhar} \code{Jalal Saadat} \code{Jamal Saadallah} \code{James} \code{Jamie} \code{Jamil Salem} \code{Jamshed} \code{Jarallah} \code{Javaid} \code{Javed} \code{Jawaid} \code{Jean Tony} \code{Jeffrey} \code{Jeffrey Leon} \code{Jehad Serwan} \code{John Earl} \code{John Walker} \code{Jorge De Los Reyes} \code{Jose} \code{Jose Guillermo} \code{Jose Tito Libio} \code{Jude Kenan} \code{Julio Cesar} \code{Juvenal Ovidio Ricardo Palmera} \code{Kamal} \code{Kamal Said} \code{Kamel} \code{Kamel Mohamed} \code{Kamran} \code{Kareem} \code{Karim} \code{Karunakaran} \code{Kenys} \code{Kevin} \code{Khaled T.} \code{Khaleel} \code{Khalid} \code{Khalid Ali-M} \code{Khalid Mohamud} \code{Khwaja Mahmood} \code{Kifah} \code{Kobie Diallo} \code{Kumeit} \code{Laguerre} \code{Lamont} \code{Latif Kamel} \code{Levar} \code{Libardo} \code{Lina} \code{Lori} \code{Luis} \code{Luis Alfredo Daza} \code{Luis Felipe Moreno} \code{Lynn} \code{Lynne} \code{Maher Mofeid} \code{Maher Yousef} \code{Mahmoud Youssef} \code{Mahmud Faruq} \code{Majeda} \code{Malek Mohamad} \code{Manel} \code{Manthena} \code{Maodo} \code{Mark Robert} \code{Marlon} \code{Maroune} \code{Martin Raouf} \code{Marwan Othman} \code{Masoud Ahmad} \code{Mazhar Iqbal} \code{Michael} \code{Michael Curtis} \code{Moeen Islam} \code{Mohadar} \code{Mohamad} \code{Mohamad Abdulamir} \code{Mohamad Ahmad} \code{Mohamad Ibrahim} \code{Mohamad Mustapha Ali} \code{Mohamed} \code{Mohamed Abdullahi} \code{Mohamed Ahmed} \code{Mohamed Mahmood} \code{Mohamed Osman} \code{Mohammad} \code{Mohammad Omar Aly} \code{Mohammad Salman Farooq} \code{Mohammad Zaki} \code{Mohammed} \code{Mohammed Ali Hassan} \code{Mohammed Ibrahim} \code{Mohammed Junaid} \code{Mohammed Mohsen Yahya} \code{Mohanad Shareef} \code{Moinuddeen} \code{Monasser} \code{Monir} \code{Montaser} \code{Monzer} \code{Moussa Ali} \code{Moustafa Habib} \code{Mubarek} \code{Mufid} \code{Muhamed} \code{Muhamed Nasir} \code{Muhammad Hamid} \code{Muhammad Ibrahim} \code{Muhammed Abid} \code{Mujahid Abdul} \code{Mukhtar} \code{Murugesu} \code{Mustafa} \code{Mustafa Ali} \code{Nabi} \code{Nabil} \code{Nabil Mohamad} \code{Nachimuthu} \code{Nadarasa} \code{Nadia} \code{Nadim} \code{Nageeb Abdul} \code{Naji Antoine Abi} \code{Najibullah} \code{Nancy} \code{Narseal} \code{Nasri} \code{Nasser} \code{Nemr Ali} \code{Nermine} \code{Nima Ali} \code{Noura} \code{Numan} \code{Nuradin M.} \code{October Martinique} \code{Omar} \code{Omar Abdi} \code{Omer} \code{Omer Abdi} \code{Onta} \code{Osama Musa} \code{Osama Yousef} \code{Osameh} \code{Oussama Abdullah} \code{Parvez Mehmood} \code{Patrice Lumumba} \code{Patricia} \code{Patrick} \code{Paul Jr.} \code{Pirouz} \code{Priscilla} \code{Raad} \code{Rabi} \code{Rafil} \code{Rafiq Abdus} \code{Rahmat} \code{Randall Todd} \code{Rania M. Fawaz} \code{Raza Nasir} \code{Reinhard} \code{Richard} \code{Richard David} \code{Robert} \code{Ronald} \code{Rothschild} \code{Roxanne} \code{Russell} \code{Sabah} \code{Saber} \code{Sabirhan} \code{Sabri} \code{Sadek} \code{Sadik} \code{Sahilal} \code{Sahim} \code{Said} \code{Saifullah Anjum} \code{Salah Osman} \code{Salam Ibrahim} \code{Saleh} \code{Saleh Ali} \code{Saleh Alli} \code{Saleh Eldin Ali} \code{Salim} \code{Salim Nemir} \code{Salman} \code{Sami Ahmad} \code{Sami Amin} \code{Sami Samir} \code{Samih Fadl} \code{Samir} \code{Sathajhan} \code{Sayed Abdul} \code{Seifullah} \code{Serdar} \code{Shafal} \code{Shahawar Matin} \code{Shain} \code{Shaker} \code{Sherif} \code{Shiraz Syed} \code{Shukri} \code{Sofiane} \code{Soliman} \code{Soliman Hamd} \code{Stanley Grant} \code{Suhail} \code{Syed} \code{Syed Gul Mohammad} \code{Syed Haris} \code{Syed Mustajab} \code{Tarek} \code{Tarek Abdelhamid} \code{Tareq Mousa} \code{Tarik} \code{Tariq} \code{Tariq Ur} \code{Thavaraja} \code{Thirunavukarasu} \code{Thiruthanikan} \code{Tomer} \code{Traci} \code{Troy M.} \code{Ulises} \code{Umar Farouk} \code{Umer} \code{Uwe} \code{Uzair} \code{Victor} \code{Victor Daniel} \code{Vijayashanthar} \code{Vincente Rafael} \code{Vinh Tan} \code{Waad Ramadan} \code{Wael} \code{Walid} \code{Walli} \code{Wassim I.} \code{Wathek} \code{Wesam} \code{Yahya} \code{Yasein} \code{Yaser} \code{Yasith} \code{Yassin} \code{Yaudat Mustafa} \code{Yehuda} \code{Yildirim Beyozit} \code{Yong Ki} \code{Youssef} \code{Zacarias} \code{Zachary} \code{Zakaria} \code{Zakariya} \code{Zameer} \code{Zarein} \code{Zeinab} \code{Ziyad} \code{Zubair A.} \code{Zuhaier Ben Mohammed} \code{Zuhair Hamed}}
    \item{\code{Charge.Date}}{a factor with levels \code{01/03/2003} \code{01/03/2006} \code{01/04/2002} \code{01/04/2003} \code{01/04/2007} \code{01/06/2010} \code{01/08/2004} \code{01/08/2010} \code{01/09/2004} \code{01/11/2002} \code{01/14/2002} \code{01/14/2003} \code{01/16/2002} \code{01/20/2004} \code{01/22/2003} \code{01/22/2010} \code{01/23/2003} \code{01/29/2003} \code{02/03/2003} \code{02/03/2005} \code{02/03/2010} \code{02/04/2002} \code{02/04/2003} \code{02/05/2002} \code{02/05/2003} \code{02/06/2006} \code{02/07/2002} \code{02/08/2006} \code{02/09/2005} \code{02/09/2006} \code{02/10/2005} \code{02/11/2009} \code{02/13/2002} \code{02/13/2003} \code{02/14/2007} \code{02/16/2006} \code{02/17/2005} \code{02/18/2005} \code{02/19/2003} \code{02/19/2009} \code{02/20/2002} \code{02/21/2002} \code{02/23/2011} \code{02/26/2009} \code{03/02/2010} \code{03/03/2010} \code{03/04/2010} \code{03/05/2004} \code{03/06/2002} \code{03/08/2006} \code{03/08/2007} \code{03/13/2003} \code{03/16/2004} \code{03/19/2003} \code{03/20/2003} \code{03/21/2007} \code{03/22/2001} \code{03/22/2011} \code{03/23/2006} \code{03/25/2003} \code{03/26/2003} \code{03/28/2001} \code{04/02/2007} \code{04/03/2002} \code{04/06/2005} \code{04/07/2003} \code{04/09/2002} \code{04/09/2003} \code{04/11/2007} \code{04/12/2005} \code{04/19/2010} \code{04/20/2005} \code{04/22/2004} \code{04/24/2003} \code{04/26/2004} \code{05/01/2003} \code{05/02/2003} \code{05/06/2004} \code{05/07/2007} \code{05/12/2011} \code{05/13/2004} \code{05/16/2003} \code{05/19/2005} \code{05/20/2002} \code{05/20/2005} \code{05/21/2003} \code{05/21/2004} \code{05/21/2007} \code{05/22/2002} \code{05/23/2005} \code{05/23/2007} \code{05/24/2006} \code{05/25/2011} \code{05/26/2011} \code{05/29/2007} \code{05/31/2005} \code{06/02/2009} \code{06/03/2003} \code{06/03/2004} \code{06/03/2010} \code{06/04/2010} \code{06/05/2002} \code{06/07/2010} \code{06/14/2004} \code{06/16/2004} \code{06/16/2005} \code{06/17/2010} \code{06/21/2006} \code{06/22/2006} \code{06/23/2011} \code{06/25/2003} \code{06/25/2004} \code{06/26/2003} \code{06/27/2002} \code{06/27/2005} \code{06/28/2007} \code{06/30/2011} \code{07/01/2003} \code{07/02/2010} \code{07/03/2003} \code{07/06/2010} \code{07/07/2010} \code{07/15/2003} \code{07/18/2005} \code{07/19/2006} \code{07/20/2010} \code{07/21/2002} \code{07/21/2010} \code{07/23/2004} \code{07/25/2002} \code{07/25/2006} \code{07/26/2004} \code{07/27/2004} \code{07/28/2004} \code{07/30/2002} \code{07/30/2003} \code{07/31/2006} \code{08/01/2007} \code{08/03/2006} \code{08/05/2004} \code{08/06/2004} \code{08/08/2003} \code{08/11/2005} \code{08/13/2003} \code{08/16/2004} \code{08/17/2005} \code{08/18/2004} \code{08/19/2003} \code{08/19/2008} \code{08/27/2004} \code{08/28/2002} \code{08/28/2006} \code{08/30/2007} \code{08/31/2005} \code{08/31/2006} \code{09/02/2008} \code{09/04/2002} \code{09/04/2003} \code{09/09/2005} \code{09/10/2004} \code{09/11/2002} \code{09/14/2005} \code{09/14/2006} \code{09/15/2003} \code{09/16/2002} \code{09/17/2004} \code{09/18/2001} \code{09/18/2009} \code{09/19/2001} \code{09/19/2006} \code{09/22/1999} \code{09/23/2004} \code{09/23/2009} \code{09/24/2009} \code{09/25/2001} \code{09/25/2007} \code{09/27/2001} \code{09/28/2003} \code{09/28/2006} \code{09/29/2006} \code{09/29/2010} \code{09/30/2002} \code{10/01/2003} \code{10/01/2009} \code{10/03/2001} \code{10/03/2002} \code{10/03/2006} \code{10/04/2001} \code{10/05/2005} \code{10/07/2004} \code{10/07/2009} \code{10/09/2001} \code{10/09/2002} \code{10/09/2003} \code{10/09/2009} \code{10/10/2001} \code{10/11/2009} \code{10/12/2001} \code{10/13/2004} \code{10/13/2009} \code{10/14/2010} \code{10/16/2001} \code{10/16/2002} \code{10/17/2001} \code{10/18/2001} \code{10/19/2009} \code{10/21/2002} \code{10/21/2010} \code{10/22/2001} \code{10/23/2001} \code{10/25/2001} \code{10/26/2001} \code{10/26/2010} \code{10/27/2009} \code{10/29/2001} \code{10/30/2001} \code{10/30/2002} \code{10/31/2001} \code{10/31/2007} \code{11/01/2001} \code{11/02/2001} \code{11/06/2001} \code{11/07/2005} \code{11/07/2008} \code{11/09/2005} \code{11/10/2004} \code{11/12/2002} \code{11/12/2010} \code{11/13/2001} \code{11/14/2001} \code{11/14/2008} \code{11/15/2001} \code{11/15/2002} \code{11/15/2006} \code{11/16/2001} \code{11/17/2005} \code{11/17/2009} \code{11/19/2001} \code{11/19/2003} \code{11/22/2006} \code{11/24/2009} \code{11/24/2010} \code{11/26/2010} \code{11/28/2001} \code{11/28/2006} \code{11/29/2001} \code{11/30/2001} \code{12/01/2004} \code{12/02/2003} \code{12/03/2003} \code{12/04/2001} \code{12/04/2002} \code{12/06/2001} \code{12/06/2010} \code{12/08/2010} \code{12/09/2004} \code{12/11/2001} \code{12/12/2001} \code{12/13/2007} \code{12/16/2003} \code{12/17/2001} \code{12/17/2002} \code{12/18/2003} \code{12/19/2001} \code{12/19/2003} \code{12/20/2006} \code{12/21/2001}}
    \item{\code{Investigation.Type.s.}}{a factor with levels \code{} \code{Agent Provocateur, Informant} \code{Agent Provocateur, Informant, Sting} \code{Agent Provocateur, Sting} \code{Informant} \code{Informant, Sting} \code{Sting}}
    \item{\code{Charges}}{a factor with levels \code{Air piracy, Air safety violations, Explosives, Murder/attempted murder of US officials, Terrorist conspiracy, Weapons of mass destruction} \code{Air piracy, Air safety violations, Funding terrorists, Material support for terrorists, Threats against the president} \code{Air safety violations} \code{Air safety violations, Attempted murder, Firearms violations, Weapons of mass destruction} \code{Air safety violations, Bombings, Explosives, Terrorist attacks on mass transport} \code{Air safety violations, Conspiracy to kill Americans, Firearms violations, Terrorist attacks on mass transport, Weapons of mass destruction} \code{Air safety violations, Immigration violations, Making false statements} \code{Arms control violations, Criminal conspiracy, Firearms violations, Material support for terrorists, RICO/racketeering} \code{Arms control violations, Criminal conspiracy, RICO/racketeering} \code{Arms control violations, Drug violations, Firearms violations, Material support for terrorists} \code{Arms control violations, Material support for terrorists, RICO/racketeering, Smuggling} \code{Asset forfeiture, Fraud/financial crimes} \code{Bombings, Conspiracy to murder, kidnap, or maim overseas, Killing Americans abroad, Material support for terrorists} \code{Bombings, Explosives, Terrorist attacks on mass transport} \code{Bribery, Contraband tobacco, Criminal conspiracy, Fraud/financial crimes, Immigration violations, Material support for terrorists, RICO/racketeering, Stolen goods/counterfeits} \code{Chemical weapons, Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Explosives, Material support for terrorists} \code{Conspiracy to kill Americans, Conspiracy to murder, Explosives, Making false statements, Murder, Murder of foreign officials, Murder/attempted murder of US officials, Weapons of mass destruction} \code{Conspiracy to kill Americans, Conspiracy to murder, kidnap, or maim overseas, Material support for terrorists, Receiving terrorist training} \code{Conspiracy to kill Americans, Conspiracy to murder, Material support for terrorists, Murder/attempted murder of US officials, Obtaining missiles to destroy airplanes} \code{Conspiracy to kill Americans, Conspiracy to murder, Material support for terrorists, Murder/attempted murder of US officials, Obtaining missiles to destroy airplanes, RICO/racketeering} \code{Conspiracy to kill Americans, Conspiracy to murder, Military sabotage, Murder/attempted murder of US officials} \code{Conspiracy to kill Americans, Explosives, Firearms violations, Funding terrorists, Material support for terrorists} \code{Conspiracy to kill Americans, Firearms violations, Murder/attempted murder of US officials} \code{Conspiracy to kill Americans, Firearms violations, Weapons of mass destruction} \code{Conspiracy to kill Americans, Killing Americans abroad, Material support for terrorists, Obtaining missiles to destroy airplanes, Weapons of mass destruction} \code{Conspiracy to murder, Conspiracy to murder, kidnap, or maim overseas, Firearms violations, Making false statements, Material support for terrorists} \code{Conspiracy to murder, Firearms violations, Murder/attempted murder of US officials} \code{Conspiracy to murder, Firearms violations, Murder/attempted murder of US officials, Weapons of mass destruction} \code{Conspiracy to murder, Firearms violations, RICO/racketeering, Sedititious conspiracy} \code{Conspiracy to murder, kidnap, or maim overseas} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Criminal solicitation, Making false statements, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Drug violations, Firearms violations, Making false statements, Material support for terrorists, Obstruction of justice, Perjury} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Fraud/financial crimes, Material support for terrorists, Terrorist fundraising} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Immigration violations, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Immigration violations, Material support for terrorists, Obstruction of justice, RICO/racketeering} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Making false statements, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Material support for terrorists, RICO/racketeering} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Weapons of mass destruction} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal solicitation, Firearms violations, Immigration violations, Making false statements, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Explosives, Making false statements, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Explosives, Material support for terrorists, Threats against the president} \code{Conspiracy to murder, kidnap, or maim overseas, Firearms violations, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Firearms violations, Material support for terrorists, Receiving terrorist training, Terrorist conspiracy, Weapons of mass destruction} \code{Conspiracy to murder, kidnap, or maim overseas, Immigration violations, Making false statements, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Making false statements, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Material support for terrorists, Terrorist conspiracy} \code{Conspiracy to murder, kidnap, or maim overseas, Material support for terrorists, Weapons of mass destruction} \code{Conspiracy to murder, Murder/attempted murder of US officials} \code{Conspiracy to murder, Murder/attempted murder of US officials, Obtaining missiles to destroy airplanes, Weapons of mass destruction} \code{Contraband tobacco, Criminal conspiracy, Drug violations, Obtaining missiles to destroy airplanes, Stolen goods/counterfeits} \code{Contraband tobacco, Criminal conspiracy, Fraud/financial crimes, Immigration violations, RICO/racketeering} \code{Contraband tobacco, Criminal conspiracy, Stolen goods/counterfeits} \code{Contraband tobacco, Drug violations, RICO/racketeering} \code{Contraband tobacco, Fraud/financial crimes, Immigration violations, Material support for terrorists, RICO/racketeering} \code{Contraband tobacco, Fraud/financial crimes, Material support for terrorists, RICO/racketeering, Terrorist fundraising} \code{Contraband tobacco, Fraud/financial crimes, RICO/racketeering, Terrorist fundraising} \code{Contraband tobacco, Material support for terrorists, RICO/racketeering, Stolen goods/counterfeits} \code{Contraband tobacco, RICO/racketeering} \code{Criminal conspiracy} \code{Criminal conspiracy, Criminal solicitation, Explosives, Firearms violations, Funding terrorists, Sedititious conspiracy} \code{Criminal conspiracy, Criminal solicitation, Making false statements, Material support for terrorists} \code{Criminal conspiracy, Drug violations, Firearms violations, Material support for terrorists} \code{Criminal conspiracy, Drug violations, Fraud/financial crimes, Immigration violations} \code{Criminal conspiracy, Drug violations, Immigration violations, Making false statements, RICO/racketeering} \code{Criminal conspiracy, Drug violations, Material support for terrorists} \code{Criminal conspiracy, Explosives, Firearms violations, Funding terrorists, Material support for terrorists, Plotting expedition against friendly nation, Sedititious conspiracy} \code{Criminal conspiracy, Explosives, Firearms violations, Material support for terrorists, Plotting expedition against friendly nation} \code{Criminal conspiracy, Explosives, Firearms violations, Material support for terrorists, Receiving terrorist training} \code{Criminal conspiracy, Export fraud, Funding terrorists} \code{Criminal conspiracy, Firearms violations} \code{Criminal conspiracy, Firearms violations, Funding terrorists} \code{Criminal conspiracy, Firearms violations, Funding terrorists, Material support for terrorists, Plotting expedition against friendly nation, Sedititious conspiracy} \code{Criminal conspiracy, Firearms violations, Immigration violations, Making false statements, Plotting expedition against friendly nation} \code{Criminal conspiracy, Firearms violations, Making false statements} \code{Criminal conspiracy, Firearms violations, Making false statements, Material support for terrorists, Plotting expedition against friendly nation} \code{Criminal conspiracy, Firearms violations, Material support for terrorists, Plotting expedition against friendly nation} \code{Criminal conspiracy, Firearms violations, Material support for terrorists, RICO/racketeering} \code{Criminal conspiracy, Firearms violations, Plotting expedition against friendly nation} \code{Criminal conspiracy, Fraud/financial crimes} \code{Criminal conspiracy, Fraud/financial crimes, Immigration violations} \code{Criminal conspiracy, Fraud/financial crimes, Immigration violations, Making false statements} \code{Criminal conspiracy, Fraud/financial crimes, Immigration violations, Making false statements, RICO/racketeering, Stolen goods/counterfeits} \code{Criminal conspiracy, Fraud/financial crimes, Immigration violations, Making false statements, RICO/racketeering, Tax violations} \code{Criminal conspiracy, Fraud/financial crimes, Making false statements} \code{Criminal conspiracy, Fraud/financial crimes, Making false statements, Obstruction of justice, RICO/racketeering} \code{Criminal conspiracy, Fraud/financial crimes, Making false statements, RICO/racketeering} \code{Criminal conspiracy, Fraud/financial crimes, Material support for terrorists, RICO/racketeering} \code{Criminal conspiracy, Fraud/financial crimes, Obstruction of justice, RICO/racketeering} \code{Criminal conspiracy, Fraud/financial crimes, RICO/racketeering} \code{Criminal conspiracy, Funding terrorists, Material support for terrorists} \code{Criminal conspiracy, Hostage taking} \code{Criminal conspiracy, Immigration violations} \code{Criminal conspiracy, Immigration violations, Making false statements} \code{Criminal conspiracy, Immigration violations, Making false statements, Material support for terrorists, Stolen goods/counterfeits} \code{Criminal conspiracy, Immigration violations, Making false statements, Obstruction of justice} \code{Criminal conspiracy, Immigration violations, Material support for terrorists} \code{Criminal conspiracy, Immigration violations, Material support for terrorists, RICO/racketeering} \code{Criminal conspiracy, Immigration violations, Stolen goods/counterfeits} \code{Criminal conspiracy, Making false statements} \code{Criminal conspiracy, Making false statements, Material support for terrorists} \code{Criminal conspiracy, Making false statements, Obstruction of justice} \code{Criminal conspiracy, Making false statements, RICO/racketeering} \code{Criminal conspiracy, Making false statements, RICO/racketeering, Stolen goods/counterfeits} \code{Criminal conspiracy, Making false statements, Tax violations} \code{Criminal conspiracy, Material support for terrorists} \code{Criminal conspiracy, Material support for terrorists, RICO/racketeering, Tax violations} \code{Criminal conspiracy, Material support for terrorists, Weapons of mass destruction} \code{Criminal conspiracy, Obstruction of justice, RICO/racketeering} \code{Criminal conspiracy, RICO/racketeering} \code{Criminal conspiracy, RICO/racketeering, Stolen goods/counterfeits} \code{Criminal conspiracy, Smuggling, Stolen goods/counterfeits} \code{Criminal conspiracy, Stolen goods/counterfeits} \code{Criminal conspiracy, Stolen goods/counterfeits, Tax violations} \code{Criminal solicitation, Explosives, Firearms violations, Material support for terrorists} \code{Drug violations} \code{Drug violations, Firearms violations, Terrorist conspiracy} \code{Drug violations, Funding terrorists} \code{Drug violations, Funding terrorists, Material support for terrorists} \code{Drug violations, Funding terrorists, Material support for terrorists, Obtaining missiles to destroy airplanes} \code{Drug violations, Immigration violations} \code{Drug violations, Making false statements} \code{Drug violations, Material support for terrorists} \code{Explosives} \code{Explosives, Firearms violations, Material support for terrorists} \code{Explosives, Firearms violations, Terrorist conspiracy, Weapons of mass destruction} \code{Explosives, Fraud/financial crimes, Material support for terrorists, Perjury, RICO/racketeering} \code{Explosives, Killing Americans abroad} \code{Explosives, Making false statements} \code{Explosives, Material support for terrorists} \code{Explosives, Material support for terrorists, Obtaining missiles to destroy airplanes} \code{Explosives, Material support for terrorists, Sedititious conspiracy} \code{Explosives, Weapons of mass destruction} \code{Extortion} \code{Firearms violations} \code{Firearms violations, Funding terrorists, Material support for terrorists, RICO/racketeering, Sedititious conspiracy} \code{Firearms violations, Funding terrorists, Material support for terrorists, Sedititious conspiracy} \code{Firearms violations, Immigration violations} \code{Firearms violations, Making false statements} \code{Firearms violations, Stolen goods/counterfeits} \code{Fraud/financial crimes} \code{Fraud/financial crimes, Funding terrorists, Immigration violations, Making false statements, RICO/racketeering, Tax violations} \code{Fraud/financial crimes, Funding terrorists, Immigration violations, RICO/racketeering} \code{Fraud/financial crimes, Immigration violations} \code{Fraud/financial crimes, Immigration violations, Making false statements} \code{Fraud/financial crimes, Immigration violations, Material support for terrorists} \code{Fraud/financial crimes, Making false statements} \code{Fraud/financial crimes, Making false statements, Material support for terrorists, RICO/racketeering, Stolen goods/counterfeits} \code{Fraud/financial crimes, RICO/racketeering, Stolen goods/counterfeits} \code{Fraud/financial crimes, Stolen goods/counterfeits} \code{Fraud/financial crimes, Tax violations} \code{Funding terrorists} \code{Funding terrorists, Immigration violations, Material support for terrorists} \code{Funding terrorists, Making false statements, Material support for terrorists} \code{Funding terrorists, Material support for terrorists} \code{Funding terrorists, Material support for terrorists, RICO/racketeering, Sedititious conspiracy} \code{Funding terrorists, Material support for terrorists, Sedititious conspiracy} \code{Hostage taking, Material support for terrorists} \code{Immigration violations} \code{Immigration violations, Making false statements} \code{Immigration violations, Making false statements, Material support for terrorists, RICO/racketeering} \code{Immigration violations, Making false statements, Obtaining sensitive defense information} \code{Immigration violations, Making false statements, Perjury} \code{Immigration violations, Material support for terrorists} \code{Immigration violations, RICO/racketeering} \code{Killing Americans abroad, Murder/attempted murder of US officials} \code{Making false statements} \code{Making false statements, Material support for terrorists} \code{Making false statements, Obstruction of justice} \code{Making false statements, Obstruction of justice, Perjury} \code{Material support for terrorists} \code{Material support for terrorists, Obstruction of justice, RICO/racketeering} \code{Material support for terrorists, Obtaining missiles to destroy airplanes} \code{Material support for terrorists, Obtaining sensitive defense information} \code{Material support for terrorists, RICO/racketeering} \code{Material support for terrorists, Terrorist attacks on mass transport} \code{Murder/attempted murder of US officials, Weapons of mass destruction} \code{Obscenity} \code{Obstruction of justice, RICO/racketeering} \code{Obtaining missiles to destroy airplanes, Weapons of mass destruction} \code{Perjury} \code{Receiving terrorist training} \code{RICO/racketeering} \code{Stolen goods/counterfeits} \code{Tax violations} \code{Terrorist conspiracy} \code{Violating trade restrictions} \code{Weapons of mass destruction}}
    \item{\code{Outcome}}{a factor with levels \code{Case pending} \code{Deported} \code{Found guilty} \code{Pleaded guilty}}
    \item{\code{Convictions}}{a factor with levels \code{} \code{Accessory after the fact} \code{Air piracy, Air safety violations, Explosives, Murder/attempted murder of US officials, Terrorist conspiracy, Weapons of mass destruction} \code{Air piracy, Air safety violations, Funding terrorists, Material support for terrorists, Threats against the president} \code{Air safety violations} \code{Air safety violations, Bombings, Explosives, Terrorist attacks on mass transport} \code{Air safety violations, Conspiracy to kill Americans, Firearms violations, Weapons of mass destruction} \code{Arms control violations, Material support for terrorists} \code{Arms control violations, Material support for terrorists, RICO/racketeering} \code{Arms control violations, Material support for terrorists, RICO/racketeering, Smuggling} \code{Arms control violations, RICO/racketeering} \code{Asset forfeiture} \code{Bombings, Conspiracy to murder, kidnap, or maim overseas, Killing Americans abroad, Material support for terrorists} \code{Bombings, Explosives, Terrorist attacks on mass transport} \code{Conspiracy to kill Americans} \code{Conspiracy to kill Americans, Conspiracy to murder, Explosives, Making false statements, Murder, Murder of foreign officials, Murder/attempted murder of US officials, Weapons of mass destruction} \code{Conspiracy to kill Americans, Conspiracy to murder, Material support for terrorists, Murder/attempted murder of US officials, Obtaining missiles to destroy airplanes, RICO/racketeering} \code{Conspiracy to kill Americans, Material support for terrorists, Receiving terrorist training} \code{Conspiracy to murder} \code{Conspiracy to murder, Firearms violations, Murder/attempted murder of US officials} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Criminal solicitation, Making false statements, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Explosives, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Criminal conspiracy, Weapons of mass destruction} \code{Conspiracy to murder, kidnap, or maim overseas, Explosives, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Immigration violations, Making false statements, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Material support for terrorists} \code{Conspiracy to murder, kidnap, or maim overseas, Material support for terrorists, Weapons of mass destruction} \code{Conspiracy to murder, Material support for terrorists, Murder/attempted murder of US officials, Obtaining missiles to destroy airplanes} \code{Conspiracy to murder, Murder/attempted murder of US officials} \code{Contraband tobacco} \code{Contraband tobacco, Criminal conspiracy, Drug violations, Obtaining missiles to destroy airplanes, Stolen goods/counterfeits} \code{Contraband tobacco, Criminal conspiracy, Immigration violations, RICO/racketeering} \code{Contraband tobacco, Fraud/financial crimes, Immigration violations, Material support for terrorists, RICO/racketeering} \code{Contraband tobacco, RICO/racketeering} \code{Criminal conspiracy} \code{Criminal conspiracy, Criminal solicitation, Explosives, Firearms violations, Funding terrorists, Sedititious conspiracy} \code{Criminal conspiracy, Criminal solicitation, Making false statements, Material support for terrorists} \code{Criminal conspiracy, Drug violations, Fraud/financial crimes, Immigration violations} \code{Criminal conspiracy, Firearms violations} \code{Criminal conspiracy, Firearms violations, Funding terrorists, Material support for terrorists, Sedititious conspiracy} \code{Criminal conspiracy, Firearms violations, Material support for terrorists} \code{Criminal conspiracy, Fraud/financial crimes} \code{Criminal conspiracy, Fraud/financial crimes, Immigration violations, Making false statements, RICO/racketeering, Stolen goods/counterfeits} \code{Criminal conspiracy, Fraud/financial crimes, Immigration violations, Making false statements, RICO/racketeering, Tax violations} \code{Criminal conspiracy, Fraud/financial crimes, Making false statements} \code{Criminal conspiracy, Fraud/financial crimes, Making false statements, RICO/racketeering} \code{Criminal conspiracy, Fraud/financial crimes, Obstruction of justice, RICO/racketeering} \code{Criminal conspiracy, Fraud/financial crimes, RICO/racketeering} \code{Criminal conspiracy, Funding terrorists, Material support for terrorists} \code{Criminal conspiracy, Hostage taking} \code{Criminal conspiracy, Immigration violations} \code{Criminal conspiracy, Immigration violations, Making false statements, Obstruction of justice} \code{Criminal conspiracy, Making false statements} \code{Criminal conspiracy, Making false statements, Obstruction of justice} \code{Criminal conspiracy, Making false statements, RICO/racketeering} \code{Criminal conspiracy, Making false statements, RICO/racketeering, Stolen goods/counterfeits} \code{Criminal conspiracy, Making false statements, Tax violations} \code{Criminal conspiracy, Material support for terrorists} \code{Criminal conspiracy, Material support for terrorists, RICO/racketeering} \code{Criminal conspiracy, Material support for terrorists, RICO/racketeering, Tax violations} \code{Criminal conspiracy, Material support for terrorists, Stolen goods/counterfeits} \code{Criminal conspiracy, Perjury, RICO/racketeering} \code{Criminal conspiracy, Smuggling} \code{Criminal conspiracy, Stolen goods/counterfeits} \code{Criminal conspiracy, Tax violations} \code{Criminal conspiracy, Terrorist fundraising} \code{Criminal solicitation, Explosives, Firearms violations, Material support for terrorists} \code{Criminal solicitation, Extortion, Material support for terrorists} \code{Drug violations} \code{Drug violations, Immigration violations} \code{Drug violations, Material support for terrorists} \code{Explosives} \code{Explosives, Firearms violations} \code{Explosives, Firearms violations, Material support for terrorists} \code{Explosives, Funding terrorists} \code{Explosives, Material support for terrorists} \code{Explosives, Material support for terrorists, Sedititious conspiracy} \code{Firearms violations} \code{Firearms violations, Funding terrorists} \code{Firearms violations, Material support for terrorists} \code{Firearms violations, Murder/attempted murder of US officials, Terrorist acts} \code{Firearms violations, Sedititious conspiracy} \code{Fraud/financial crimes} \code{Fraud/financial crimes, Funding terrorists, Immigration violations, RICO/racketeering} \code{Fraud/financial crimes, Immigration violations} \code{Fraud/financial crimes, Immigration violations, Making false statements} \code{Fraud/financial crimes, Making false statements} \code{Fraud/financial crimes, Material support for terrorists, RICO/racketeering} \code{Fraud/financial crimes, RICO/racketeering} \code{Funding terrorists} \code{Funding terrorists, Immigration violations, Material support for terrorists} \code{Funding terrorists, Immigration violations, Tax violations} \code{Funding terrorists, Material support for terrorists} \code{Hostage taking} \code{Immigration violations} \code{Immigration violations, Making false statements} \code{Immigration violations, Making false statements, Obtaining sensitive defense information} \code{Killing Americans abroad} \code{Making false statements} \code{Making false statements, Material support for terrorists} \code{Making false statements, Material support for terrorists, RICO/racketeering} \code{Making false statements, Obstruction of justice, Perjury} \code{Material support for terrorists} \code{Material support for terrorists, Obtaining missiles to destroy airplanes} \code{Material support for terrorists, Obtaining sensitive defense information} \code{Material support for terrorists, RICO/racketeering} \code{Material support for terrorists, Terrorist attacks on mass transport} \code{Military sabotage} \code{Murder/attempted murder of US officials, Weapons of mass destruction} \code{Obscenity} \code{Obstruction of justice} \code{Obtaining missiles to destroy airplanes, Weapons of mass destruction} \code{Perjury} \code{Receiving terrorist training} \code{RICO/racketeering} \code{RICO/racketeering, Terrorist fundraising} \code{Sedititious conspiracy} \code{Stolen goods/counterfeits} \code{Tax violations} \code{Violating trade restrictions} \code{Weapons of mass destruction}}
    \item{\code{State}}{a factor with levels \code{Alabama} \code{Alaska} \code{Arizona} \code{Arkansas} \code{California} \code{Colorado} \code{Connecticut} \code{Delaware} \code{District of Columbia} \code{Florida} \code{Georgia} \code{Illinois} \code{Indiana} \code{Iowa} \code{Kentucky} \code{Louisiana} \code{Maine} \code{Maryland} \code{Massachusetts} \code{Michigan} \code{Minnesota} \code{Mississippi} \code{Missouri} \code{Montana} \code{New Jersey} \code{New York} \code{North Carolina} \code{Ohio} \code{Oklahoma} \code{Oregon} \code{Pennsylvania} \code{South Carolina} \code{Tennessee} \code{Texas} \code{Virginia} \code{Washington} \code{Wisconsin}}
    \item{\code{Alleged.Affiliation}}{a factor with levels \code{} \code{Abu Sayyaf} \code{al-Aqsa Martyrs' Brigades} \code{al-Barakat} \code{al-Fuqra} \code{al-Haramain Islamic Foundation} \code{al-Ittihad al-Islami} \code{al-Shabaab} \code{Al Gamaa al-Islamiyya} \code{Al Qaeda} \code{Al Qaeda, al-Shabaab} \code{Al Qaeda, Hamas} \code{Ansar al-Islam} \code{Free Government of Vietnam} \code{Hamas} \code{Hezbollah} \code{Iraqi insurgency} \code{Jam'iyyat Ul-Islam Is-Saheeh} \code{Jemaah Islamiyah} \code{Khalistan Commando Force (KCF)} \code{Lashkar-e-Taiba} \code{Mujahideen-e-Khalq} \code{Palestinian Islamic Jihad} \code{Revolutionary Armed Forces of Colombia (FARC)} \code{Taliban} \code{Tamil Tigers} \code{Terrorist training camps in Afghanistan and Pakistan} \code{United Self-Defense Forces of Colombia (AUC)}}
    \item{\code{month}}{a numeric vector}
    \item{\code{year}}{a numeric vector}
    \item{\code{day}}{a numeric vector}
    \item{\code{BCriminal.conspiracy}}{a numeric vector}
    \item{\code{BFirearms.violations}}{a numeric vector}
    \item{\code{BImmigration.violations}}{a numeric vector}
    \item{\code{BMaking.false.statements}}{a numeric vector}
    \item{\code{BPlotting.expedition.against.friendly.nation}}{a numeric vector}
    \item{\code{BContraband.tobacco}}{a numeric vector}
    \item{\code{BStolen.goods.counterfeits}}{a numeric vector}
    \item{\code{BFunding.terrorists}}{a numeric vector}
    \item{\code{BMaterial.support.for.terrorists}}{a numeric vector}
    \item{\code{BConspiracy.to.murder..kidnap..or.maim.overseas}}{a numeric vector}
    \item{\code{BFraud.financial.crimes}}{a numeric vector}
    \item{\code{BCriminal.solicitation}}{a numeric vector}
    \item{\code{BConspiracy.to.murder}}{a numeric vector}
    \item{\code{BMurder.attempted.murder.of.US.officials}}{a numeric vector}
    \item{\code{BWeapons.of.mass.destruction}}{a numeric vector}
    \item{\code{BAir.safety.violations}}{a numeric vector}
    \item{\code{BAttempted.murder}}{a numeric vector}
    \item{\code{BRICO.racketeering}}{a numeric vector}
    \item{\code{BDrug.violations}}{a numeric vector}
    \item{\code{BExplosives}}{a numeric vector}
    \item{\code{BSedititious.conspiracy}}{a numeric vector}
    \item{\code{BAir.piracy}}{a numeric vector}
    \item{\code{BThreats.against.the.president}}{a numeric vector}
    \item{\code{BTax.violations}}{a numeric vector}
    \item{\code{BObtaining.sensitive.defense.information}}{a numeric vector}
    \item{\code{BPerjury}}{a numeric vector}
    \item{\code{BTerrorist.conspiracy}}{a numeric vector}
    \item{\code{BTerrorist.attacks.on.mass.transport}}{a numeric vector}
    \item{\code{BObstruction.of.justice}}{a numeric vector}
    \item{\code{BConspiracy.to.kill.Americans}}{a numeric vector}
    \item{\code{BObtaining.missiles.to.destroy.airplanes}}{a numeric vector}
    \item{\code{BMilitary.sabotage}}{a numeric vector}
    \item{\code{BTerrorist.fundraising}}{a numeric vector}
    \item{\code{BKilling.Americans.abroad}}{a numeric vector}
    \item{\code{BArms.control.violations}}{a numeric vector}
    \item{\code{BBombings}}{a numeric vector}
    \item{\code{BReceiving.terrorist.training}}{a numeric vector}
    \item{\code{BAsset.forfeiture}}{a numeric vector}
    \item{\code{BExport.fraud}}{a numeric vector}
    \item{\code{BBribery}}{a numeric vector}
    \item{\code{BMurder}}{a numeric vector}
    \item{\code{BMurder.of.foreign.officials}}{a numeric vector}
    \item{\code{BChemical.weapons}}{a numeric vector}
    \item{\code{BObscenity}}{a numeric vector}
    \item{\code{BSmuggling}}{a numeric vector}
    \item{\code{BHostage.taking}}{a numeric vector}
    \item{\code{BViolating.trade.restrictions}}{a numeric vector}
    \item{\code{BExtortion}}{a numeric vector}
    \item{\code{CFirearms.violations}}{a numeric vector}
    \item{\code{CFunding.terrorists}}{a numeric vector}
    \item{\code{CMaking.false.statements}}{a numeric vector}
    \item{\code{CContraband.tobacco}}{a numeric vector}
    \item{\code{CMaterial.support.for.terrorists}}{a numeric vector}
    \item{\code{CStolen.goods.counterfeits}}{a numeric vector}
    \item{\code{CCriminal.conspiracy}}{a numeric vector}
    \item{\code{CFraud.financial.crimes}}{a numeric vector}
    \item{\code{CRICO.racketeering}}{a numeric vector}
    \item{\code{CDrug.violations}}{a numeric vector}
    \item{\code{CExplosives}}{a numeric vector}
    \item{\code{CAir.piracy}}{a numeric vector}
    \item{\code{CAir.safety.violations}}{a numeric vector}
    \item{\code{CThreats.against.the.president}}{a numeric vector}
    \item{\code{CTax.violations}}{a numeric vector}
    \item{\code{CObtaining.sensitive.defense.information}}{a numeric vector}
    \item{\code{CImmigration.violations}}{a numeric vector}
    \item{\code{CConspiracy.to.murder..kidnap..or.maim.overseas}}{a numeric vector}
    \item{\code{CTerrorist.attacks.on.mass.transport}}{a numeric vector}
    \item{\code{CWeapons.of.mass.destruction}}{a numeric vector}
    \item{\code{CConspiracy.to.kill.Americans}}{a numeric vector}
    \item{\code{CConspiracy.to.murder}}{a numeric vector}
    \item{\code{CMurder.attempted.murder.of.US.officials}}{a numeric vector}
    \item{\code{CObtaining.missiles.to.destroy.airplanes}}{a numeric vector}
    \item{\code{CMilitary.sabotage}}{a numeric vector}
    \item{\code{CCriminal.solicitation}}{a numeric vector}
    \item{\code{CSedititious.conspiracy}}{a numeric vector}
    \item{\code{CPerjury}}{a numeric vector}
    \item{\code{CTerrorist.fundraising}}{a numeric vector}
    \item{\code{CObstruction.of.justice}}{a numeric vector}
    \item{\code{CExtortion}}{a numeric vector}
    \item{\code{CBombings}}{a numeric vector}
    \item{\code{CKilling.Americans.abroad}}{a numeric vector}
    \item{\code{CAsset.forfeiture}}{a numeric vector}
    \item{\code{CArms.control.violations}}{a numeric vector}
    \item{\code{CMurder}}{a numeric vector}
    \item{\code{CMurder.of.foreign.officials}}{a numeric vector}
    \item{\code{CAccessory.after.the.fact}}{a numeric vector}
    \item{\code{CObscenity}}{a numeric vector}
    \item{\code{CSmuggling}}{a numeric vector}
    \item{\code{CReceiving.terrorist.training}}{a numeric vector}
    \item{\code{CTerrorist.conspiracy}}{a numeric vector}
    \item{\code{CHostage.taking}}{a numeric vector}
    \item{\code{CTerrorist.acts}}{a numeric vector}
    \item{\code{CViolating.trade.restrictions}}{a numeric vector}
    \item{\code{nrCharges}}{a numeric vector}
    \item{\code{nrConvs}}{a numeric vector}
  }
}
\examples{
data(MJnew)
## maybe str(MJnew) ; plot(MJnew) ...
}
\keyword{datasets}
