\name{olives}
\alias{olives}
\docType{data}
\title{
olive oil fatty acids
}
\description{
Various fatty acid measurements.
}
\usage{data(olives)}
\format{
  A data frame with 572 observations on the following 11 variables.
  \describe{
    \item{\code{Area}}{a factor with levels \code{Calabria} \code{Coast-Sardinia} \code{East-Liguria} \code{Inland-Sardinia} \code{North-Apulia} \code{Sicily} \code{South-Apulia} \code{Umbria} \code{West-Liguria}}
    \item{\code{Region}}{a factor with levels \code{North} \code{Sardinia} \code{South}}
    \item{\code{palmitic}}{a numeric vector}
    \item{\code{palmitoleic}}{a numeric vector}
    \item{\code{stearic}}{a numeric vector}
    \item{\code{oleic}}{a numeric vector}
    \item{\code{linoleic}}{a numeric vector}
    \item{\code{linolenic}}{a numeric vector}
    \item{\code{arachidic}}{a numeric vector}
    \item{\code{eicosenoic}}{a numeric vector}
    \item{\code{Test.Training}}{a factor with levels \code{Test} \code{Training}}
  }
}

\examples{
data(olives)
}
\keyword{datasets}
