\name{plants}
\alias{plants}
\docType{data}
\title{
Plants
}
\description{
Binary state variables indicating which of more than 30000 plants grow in that state.
}
\usage{data(plants)}
\format{
  A data frame with 34781 observations on the following 70 variables.
  \describe{
    \item{\code{V1}}{name}
    \item{\code{ab}}{a numeric vector}
    \item{\code{ak}}{a numeric vector}
    \item{\code{ar}}{a numeric vector}
    \item{\code{az}}{a numeric vector}
    \item{\code{ca}}{a numeric vector}
    \item{\code{co}}{a numeric vector}
    \item{\code{ct}}{a numeric vector}
    \item{\code{de}}{a numeric vector}
    \item{\code{dc}}{a numeric vector}
    \item{\code{fl}}{a numeric vector}
    \item{\code{ga}}{a numeric vector}
    \item{\code{hi}}{a numeric vector}
    \item{\code{id}}{a numeric vector}
    \item{\code{il}}{a numeric vector}
    \item{\code{in.}}{a numeric vector}
    \item{\code{ia}}{a numeric vector}
    \item{\code{ks}}{a numeric vector}
    \item{\code{ky}}{a numeric vector}
    \item{\code{la}}{a numeric vector}
    \item{\code{me}}{a numeric vector}
    \item{\code{md}}{a numeric vector}
    \item{\code{ma}}{a numeric vector}
    \item{\code{mi}}{a numeric vector}
    \item{\code{mn}}{a numeric vector}
    \item{\code{ms}}{a numeric vector}
    \item{\code{mo}}{a numeric vector}
    \item{\code{mt}}{a numeric vector}
    \item{\code{ne}}{a numeric vector}
    \item{\code{nv}}{a numeric vector}
    \item{\code{nh}}{a numeric vector}
    \item{\code{nj}}{a numeric vector}
    \item{\code{nm}}{a numeric vector}
    \item{\code{ny}}{a numeric vector}
    \item{\code{nc}}{a numeric vector}
    \item{\code{nd}}{a numeric vector}
    \item{\code{oh}}{a numeric vector}
    \item{\code{ok}}{a numeric vector}
    \item{\code{or}}{a numeric vector}
    \item{\code{pa}}{a numeric vector}
    \item{\code{pr}}{a numeric vector}
    \item{\code{ri}}{a numeric vector}
    \item{\code{sc}}{a numeric vector}
    \item{\code{sd}}{a numeric vector}
    \item{\code{tn}}{a numeric vector}
    \item{\code{tx}}{a numeric vector}
    \item{\code{ut}}{a numeric vector}
    \item{\code{vt}}{a numeric vector}
    \item{\code{va}}{a numeric vector}
    \item{\code{vi}}{a numeric vector}
    \item{\code{wa}}{a numeric vector}
    \item{\code{wv}}{a numeric vector}
    \item{\code{wi}}{a numeric vector}
    \item{\code{wy}}{a numeric vector}
    \item{\code{al}}{a numeric vector}
    \item{\code{bc}}{a numeric vector}
    \item{\code{mb}}{a numeric vector}
    \item{\code{nb}}{a numeric vector}
    \item{\code{lb}}{a numeric vector}
    \item{\code{nf}}{a numeric vector}
    \item{\code{nt}}{a numeric vector}
    \item{\code{ns}}{a numeric vector}
    \item{\code{nu}}{a numeric vector}
    \item{\code{on}}{a numeric vector}
    \item{\code{pe}}{a numeric vector}
    \item{\code{qc}}{a numeric vector}
    \item{\code{sk}}{a numeric vector}
    \item{\code{yt}}{a numeric vector}
    \item{\code{dengl}}{a numeric vector}
    \item{\code{fraspm}}{a numeric vector}
  }
}
\source{
formerly: archive.ics.uci.edu/ml/datasets/Plants
}
\keyword{datasets}
