\name{extracat}
\alias{extracat}

\docType{package}
\title{
Graphical extensions for catagorical data
}
\description{
Provides interactive parallel coordinates plots based on package \code{iplots} for categorical and continuous data
as well as a grid-based special version of multiple barcharts which shows the relative frequencies of a target variable an the corresponding weights in a MB-like graphical layout.
The second plot can also be used as a generalization of spineplots.
The main functions are \link[extracat]{cpcp} and \link[extracat]{rmb}.
The interactive version of the latter (\link[extracat]{irmb}) is based on \pkg{iWidgets} and thus also on \pkg{JGR}.
}
\details{
\tabular{ll}{
Package: \tab extracat\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2010-02-23\cr
License: \tab -\cr
LazyLoad: \tab yes\cr
}

}
\author{
Alexander Pilhoefer\cr
Department for Computer Oriented Statistics and Data Analysis\cr
University of Augsburg\cr
Germany

Maintainer: Alexander Pilhoefer \email{alexander.pilhoefer@math.uni-augsburg.de}
}
\references{
Alexander Pilhoefer  \emph{Analysen von kategoriellen Daten}\cr Diploma Thesis 2009\cr
\cr
  Alexander Pilhoefer \emph{ New approaches in visualization of categorical data:
       R-package extracat}\cr
  Journal of Statistical Software, submitted March 2010

}

\keyword{ categorical }
\keyword{ parallel coordinates }
\keyword{ multiple barcharts }
\keyword{ mosaicplots }
\keyword{ interactive }
\seealso{
}
\examples{

}
